Add ViewMsg_DoneUsingBitmap message.
ViewMsg_UpdateRect_ACK messages sent from the browser to the
renderer currently tell the renderer both that it can start
rendering the next update and that the TransportDIB that it
previously sent to the browser is now available for reuse.
This change adds a new ViewMsg_DoneUsingBitmap message for
communicating the latter piece of information.  We currently
always send ViewMsg_DoneUsingBitmap immediately before
sending ViewMsg_UpdateRect_ACK, so no functional changes are
intended from this change.

This is preparatory work for an optimization where we defer
copying updates to the backing store, instead copying them
directly from the TransportDIB to the widget -- if we get
repeated updates for the same region, we can avoid copying
all but the final update to the backing store entirely.

BUG=none
TEST=tried it on GTK; sent to trybots.  will do some testing on other platforms too

Review URL: http://codereview.chromium.org/3506007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@61225 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: b7b62e147d7bf225be76be14d005f06008f07c7e