Remove spammy warning about reaching maximum socket count for the pool.
When I run Chrome through a proxy, this WARNING log triggers very frequently. Typically immediately after it's called, CloseOneIdleSocket() is invoked and a connect job is still allowed to proceed.

If this WARNING was useful in the past, I will add it to the call sites at appropriate places (such as when there are no idle sockets for the pool and a request must be queued) rather than inside ReachedMaxSocketsLimit().

BUG=None
TEST=None

Review URL: http://codereview.chromium.org/3526013

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@61363 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 2712bd7e0f547ec60e924face859f97565d2a2bc