linux: Fix sort order in the task maanger.
In Linux, the sort ordering in the task manager is incorrect, as in when you
click on a column header to sort the values the first time, it shows a triangle
with the pointy head downwards, i.e. it should show the values from larger to
smaller. But currently, the values are shown smaller to larger. So this small
patch is a fix for that. This is also consistent with the behaviour on Windows.

Also, this makes more sense. When a user is sorting, for example, the memory or
the cpu column, it is more likely he is interested on the pages that are taking
the most memory/cpu, so they should be at the top for the first sort.

BUG=none
TEST=after opening multiple tabs, bring up the task manager, then click on the
memory column header. It should show the page using the most memory on top.

patch by Sadrul Chowdhury <sadrul [at] chromium [dot] org>
original review: http://codereview.chromium.org/3590009

Review URL: http://codereview.chromium.org/3615006

git-svn-id: http://src.chromium.org/svn/trunk/src@61387 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 531aa427a6f1ad05bbd19db6007aaaf235131e80