Add the test for the AllUrls Extension feature.
The test was failing on the bots because the extensions get different ids
on different bots, since the paths are different. I've changed the test to
load the extensions first and then query their id, instead of relying on
the id they get on my computer.

BUG=36275
TEST=AllUrlsApiTest (2 new ones)

Review URL: http://codereview.chromium.org/3610006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@61419 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 96409752c4bfb7dce93f64eb53fb0030cc8689e0