DOMUI Options: Fix vertical alignment of radio and checkboxes
This also makes the label text line up correctly if the label needs to wrap.

This also fixes a bug where the menu could be scrolled off the screen.

I went through and fixed a bunch of indentation errors in the HTML I touched.

BUG=None
TEST=None

Review URL: http://codereview.chromium.org/3590006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@61544 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 175fdf39434b401ba1423920d88d3e5a11fde86d