Add a glyph drawing routine to private Pepper interface. The current implementation is Linux only.
This also changes the Private font interface to use the "regular"
PP_FontDescription_Dev struct for specifying the font, which allows us to unify
some of the data here. I kept the character set stuff as a separate parameter
which is needed for compat in a few cases.

TEST=none
BUG=none
Review URL: http://codereview.chromium.org/3569013

git-svn-id: http://src.chromium.org/svn/trunk/src@61920 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 3b7d778ee33528fc9add46a8f4ce190aeffd3a99