The problem here seems to have been a race condition between the animator wanting to draw the status bubble and the browser window closing sequence.  If  the animation is always reset when the status bubble is detached (which is the correct thing to do, and this not being the case was actually causing another, more subtle bug), the animation sequence should never be running when the window is shutting down.
(No delegate is needed to watch for the expansion stop, because this case doesn't trigger a state change -- the state "is_expanded" is true when the bubble has been expanded, and ceases to be true when the bubble is hidden. The bubble will remain in "is_expanded" mode while the user moves quickly from one bubble to the next, expanding and contracting to fit the new links without pausing for a hover delay).

BUG=57600
TEST=open several browsers. hover over a long (elided) link on the top browser. wait for the expansion -- kill the window quickly, as it expands! repeat -- browser should not crash. browser should never crash if closing while expanding in any circumstance.
Review URL: http://codereview.chromium.org/3564021

git-svn-id: http://src.chromium.org/svn/trunk/src@61973 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 4e4919e56eedb2bf181f49d09a1ab7a624ded6e8