Cloud Print Proxy UI fixes.
When the authentication dialog is active, we have to re-layout because
the in-process text put on the button might be larger.

When setting up the desciption for enabled or disabled state, get the
containing scroll view to do the layout, because the height of the
description may have changed.  Be nice to the user and make sure the
right things are visible again after that layout.

BUG=none
TEST=use a really long username to enable the cloud print proxy such that the description text wraps to 3 lines.  The UI should stay all visible with nothing clipped.

Review URL: http://codereview.chromium.org/3701003

git-svn-id: http://src.chromium.org/svn/trunk/src@62196 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: a8528cc7a3a4d5c309f3dbc852dbe578afb43117