Reland fix for OCSP startup race.
This is a revert of r62107 which is a revert of r62105 which is a revert of r60753 which is a revert of r60739 which is a revert of r60025 which is a revert of r59972  which is a revert of r59570 which is a revert of r59511 which is a revert of r59299 which is a revert of r59289.

The most recent reverts have all been due to ChromiumOS leaks which I believe to be fixed after fixing bug 58572, since I tested the combined fix in r62105 which was green on all ChromiumOS valgrind bots.

BUG=55940
TEST=Run a debug build on Linux.  10% or so of startups used to crash on startup.  Shouldn't happen anymore.

Review URL: http://codereview.chromium.org/3660010

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@62256 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: bcbcaa3200d4e0fbf0aca1ea41dba77eb488af7d