Reland fix for OCSP startup race.
This is a revert of r62107 which is a revert of r62105 which is a revert of r60753 which is a revert of r60739 which is a revert of r60025 which is a revert of r59972  which is a revert of r59570 which is a revert of r59511 which is a revert of r59299 which is a revert of r59289.

The most recent reverts have all been due to ChromiumOS leaks which I believe to be fixed after fixing bug 58572, since I tested the combined fix in r62105 which was green on all ChromiumOS valgrind bots.

BUG=55940
TEST=Run a debug build on Linux.  10% or so of startups used to crash on startup.  Shouldn't happen anymore.

Review URL: http://codereview.chromium.org/3660010

git-svn-id: http://src.chromium.org/svn/trunk/src@62256 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: b6f0f4dfda45ba58c4ce9dcd4a112b985f9e306a