Compile webkit gpu code in an upstream build too.
I had originally disabled this when setting up the upstream
webkit.org build.  At the time we didn't compile any of src/gpu,
but now we do, so it's safe to enable this.

If we don't compile this, shared library (.so) builds in a
webkit only tree fail to link because of missing symbols.

Review URL: http://codereview.chromium.org/3722002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@62330 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: fec8ea9c3ff7894c08471fd25870f6367d3c3ad9