Fix flaky IncognitoSplitMode. For real this time.
I accidentally checked in the wrong change last time.

This reverts r62432 and commits the code that was actually reviewed in http://codereview.chromium.org/3656005/show .

Stop processing incoming messages in BrowserRenderProcessHost when it is about
to be deleted. See comment 20 on the bug for details.

BUG=53991
TEST=handled by existing test

Review URL: http://codereview.chromium.org/3772003

git-svn-id: http://src.chromium.org/svn/trunk/src@62445 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 870d0315814bedc37b3a6ac7ff87957ff2032769