Revert 62605 - Make Run methods non-virtual on RRResolverWorker.
I don't think this needs to be reverted, but git cl dcommit went a bit wacky and I'm being conservative.

The goal is to turn on -Wnon-virtual-dtor warning for gcc. In this case, the virtual Run()'s do not appear necessary,
especially since they are private. If intended to be derived from, there is a "delete this" in the base class which would
point to requiring a virtual destructor.

BUG=45135
TEST=Target builds, validated that there are no derived classes of RRResolverWorker.

Review URL: http://codereview.chromium.org/3824002

TBR=cbentzel@chromium.org
Review URL: http://codereview.chromium.org/3813005

git-svn-id: http://src.chromium.org/svn/trunk/src@62611 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 940f70d30e0a9234edcff572fb0b7ae90af7e013