This adds a plugin selection policy for selecting which plugin is allowed for a particular domain.
It is only used on ChromeOS.  It reads from a file that is installed
in a known location on ChromeOS, and uses that as it's policy.

When there are multiple plugins supporting the same mime-type, the
appropriate plugin file to load is now selected based on policy.

BUG=http://crosbug.com/7403
TEST=ran new unit test, tested on device.

Review URL: http://codereview.chromium.org/3717005

git-svn-id: http://src.chromium.org/svn/trunk/src@62679 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: b7be29710674dd8185c00c2d36f0cc75b5fbc29a