Cut down timeouts in FilePathWatcherTest.
Make the test continue immediately when a notification is received and remove all tests that check for the absence of notifications, the FilePathWatcher code may generate spurious notifications anyway. Move the tests over to browser_tests so they can run in a separate process. This avoid tinkering with the message loop to catch timeouts but just uses the built-in browser_tests timeouts.

BUG=56868
TEST=Running FilePathWatcherTest should take only a few seconds instead of up to a minute.

Review URL: http://codereview.chromium.org/3555016

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@62748 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 1ec9b9b0a3a5512b561c756ee336ad279475a054