about:flags: Fix disabling experiments.
The problem was that the restart code would copy the switches added by about:flags to the restarted instance, so that e.g. --enable-expose-for-tabs would be passed on the the command line even though the experiment was disabled. Now the flags added by about:switches are explicitly removed before the restart.

Also wrap the switches added by about:flags with --start-flags-switches and --end-flags-switches switches. These unknown switches are ignored, but are useful to see on about:version.

Also add a bunch of unrelated unit tests.

BUG=56314
TEST=See bug.

Review URL: http://codereview.chromium.org/3813007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@62777 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 27e9cc8c5efc8d90346e4fe9fe01304f143d7f6d