Tweak the webstorePrivate.getBrowserLogin callback signature.
I accidentally defined it as taking 2 parameters intstead of a single object
with two named properties. This CL fixes that (and incidentally happens to make
it match the callback for the promptBrowserLogin callback, which is nice for
consistency).

I also did a little cleanup and added a test for getBrowserLogin and added 
another test for promptBrowserLogin while I was in the neighborhood.

BUG=58033
TEST=none

Review URL: http://codereview.chromium.org/3767010

git-svn-id: http://src.chromium.org/svn/trunk/src@62805 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 5984514c2dff9e34d91b45e9bd47663ef7d24b77