net: move SSL host info serialisation into SSLHostInfo
This is a prelude to SSLHostInfo being able to kick off certificate
validation based on predicted certificates.

In order for that to happen, SSLHostInfo has to be able to parse the
contents of the saved data, therefore the
serialisation/deserialisation is moved into SSLHostInfo and we use a
protocol buffer to clean things up.

TEST=net_unittests
BUG=none

Review URL: http://codereview.chromium.org/3915001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@63221 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 8cd06081ce8b3f2ab900ec3b6ef2330876087928