Fix a failing test for the extension webstorePrivate API.
The test has been failing very frequently, always hitting the expectation
in FakeProfileSyncService::RemoveObserver that the observer being removed isn't
the currently registered one.

I think the problem is that I was waiting until the PromptBrowserLoginFunction
destructor (which is called by the ref count going to 0) to have the function
remove itself as a listener, when really once the OnStateChanged function 
fires and sync setup is complete we don't want to be listening anymore.

BUG=59419
TEST=The ExtensionWebstorePrivateBrowserTest.BrowserLogin browser test should
stop failling

Review URL: http://codereview.chromium.org/3974002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@63315 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: e14237c898107c54c0a480be99713c65ba832c78