Hijack mouse-related events for TOUCH_UI builds
With this change, which is specifically for TOUCH_UI builds, mouse-related
messages are hijacked from the normal Gdk event dispatcher and processed
manually in the file touchui/touchui.cc.  GdkEvents are converted directly
to views events, and dispatched to the RootView.  This is preliminary work
that will be followed by more elaborate message pump changes, and ultimately
is in the spirit of removing Gtk entirely for TOUCH_UI (Chromium-Views-Gtk).

Patch from Chad Faragher <wyck@chromium.org>

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/3704005

git-svn-id: http://src.chromium.org/svn/trunk/src@63368 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 8b9d3c62c1da3dbe4e085a414118e9d7ac33a850