Fix opening bookmarks in an incognito window.
This is a fix for bug 59656.

Looks like this bug was introduced by http://codereview.chromium.org/3210007
Because of this change, the callback for chrome.windows.create gets
'null' when |incognito| is true. As a result, the tabs after the
first one cannot be added to this new incognito window.

However, the API enhancement from http://codereview.chromium.org/3772004
allows creating a new window with multiple tabs. So use this API to
open the bookmarks in the new window.

Patch from Sadrul Chowdhury <sadrul@chromium.org>

BUG=59656
TEST=see bug

Review URL: http://codereview.chromium.org/4022005

git-svn-id: http://src.chromium.org/svn/trunk/src@63763 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: f38789842ba6d8c88bfd49dd6180018b78deb49e