chromeos: Avoid assert if WM tells us to select invalid tab.
The window manager gets notified asynchronously when the
number of tabs in a window changes, so there's no guarantee
that the select-tab messages that the WM sends will always
have a valid index in them.  This change makes us ignore
invalid messages instead of CHECK()-ing when we receive
them.

BUG=chromium-os:8103
TEST=built an evil window manager that always sends index -1 and checked that chrome stops crashing with this change

Review URL: http://codereview.chromium.org/4089002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@63803 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: b07f333bb2f8a376abca31509759023a5c617858