ChromeFrame network tests were broken since the change to honor the automation protocol version in ChromeFrame which basically breaks the automation connection if the protocols don't match. While this is fine when chrome is the automation client, it breaks chrome frame net tests because when the automation channel is established the automation provider sends over the version of the current module in the hello message.
In the case when the current module is chrome the version goes over correctly. For chrome frame
net tests there is no version string. Hence the channel never gets established.

Fix is to retrieve the protocol version via a virtual protected function in the automation
provider. The testing automation provider overrides this function in the network tests
and returns the version of chrome.dll

TBR=amit,robertshield

Review URL: http://codereview.chromium.org/4160002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@63860 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d5a6352ed5d71f60736c64360ec7206b5b383c03