Add logo display for fourth search engine in search engine selection dialog.
BUG=57993
TEST=Set the search engine on your (non-Chrome) default browser to be NOT in the top three for your region.
For USA, for example, you could use Baidu or Aol Search. Start Chrome with --first-run switch on an empty profile.
You should see the fourth search engine selection represented by a logo, not just text (if it's included in the list of logos for which we have images).
Review URL: http://codereview.chromium.org/4109001

git-svn-id: http://src.chromium.org/svn/trunk/src@64074 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: ff49258d2cbd8800507833d9e47177419ebf41a2