Add support for OnClose event in ActiveX and NPAPI
Provide event notification for window.close called within
ActiveX and NPAPI. Currently this is handled in ActiveDocument
to close the popups. In ActiveX and NPAPI plugin we just fire
an event as it is upto the container to decide what's the right
thing to do.

BUG=59664
TEST=New tests ChromeFrameTestWithWebServer.FullTabModeIE_WindowClose,  ChromeFrameTestWithWebServer.FullTabModeFF_WindowClose

Review URL: http://codereview.chromium.org/4152005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@64092 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: e7c1c25b90342ea3e504b37b6df03ce6bfeb595a