Changes to enable retry of failed sends on feedback reports.
If the initial send fails, the first retry will be in 15 minutes after the
attempt. The next attempt will be in 30 minutes and the time will double till
we hit a 4 hour delay - the send will be then retried every 4 hours till the
Chrome instance is up.

BUG=cros:6331
TEST=Tested with up to 25 reports at a time, with varying network connectivity
(via pulling out the network cable); report data verified on the feedback
server.

Review URL: http://codereview.chromium.org/4027005

git-svn-id: http://src.chromium.org/svn/trunk/src@64321 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: c0d2d241436522795ba6bdd9fdfbd54759a69c58