[cros] Screen locker should try to auth on the UI thread, not the FILE thread.
LoadLocalaccount can be called asynchronously in AuthenticateToUnlock.
Doing a CheckOffline() will fail over to a localaccount check anyway,
and it will appropriately ensure that the localaccount file has been
loaded before trying to check.  This means that AuthenticateToUnlock
can be called on the UI thread, as it should be.

BUG=chromium-os:8271
TEST=unit tests, browser tests, install on device and try to lock/unlock with real account and localaccount

Review URL: http://codereview.chromium.org/4106008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@64538 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 3d4eaecce458f786df304c762082763bb64d0142