Handle selection changes due to AutoFill more carefully.
When previewing a form, select the contents of the initiating field so that typing will overwrite the field's contents.
When a form is filled, move the cursor to the end of the initiating field, so that typing appends text.
When a form is cleared, make sure that we restore the original value of the initiating field.

BUG=58774
TEST=browser_tests --gtest_filter=FormManagerTest.*Fill*:FormManagerTest.*Preview*

Review URL: http://codereview.chromium.org/4221002

git-svn-id: http://src.chromium.org/svn/trunk/src@64612 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: ed1d9f8afded7051fe596f8adcc0ea4c42e36fa6