Adds error text for the ERR_NETWORK_ACCESS_DENIED error, which is generally caused by a firewall.
Also removes the title strings for proxy and disconnected
errors, replacing them with pre-existing generic ones, and
sorting them by their new title.  This avoids a debug
assertion caused by the fact neither title includes the url
of the original page.  Not much of titles are seen anyways,
so not worth the effort of a workaround.

BUG=57108
TEST=manual

Review URL: http://codereview.chromium.org/4182012

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@64786 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 5dda7a6cbf35276a8de8f80170dcc5ee74d92334