Change disabling the assert that is causing ExtensionApiTest.Popup to become flaky.
I've tried to reproduce the flaky behaviour on my local machine, in a VM, and on the build-bots, but have not been able to reproduce the problem.  Because this test is important, I am disabling the single assert so that the flaky label can be removed.

BUG=46601
TEST=ExtensionApiTest.Popup

Review URL: http://codereview.chromium.org/4264005

git-svn-id: http://src.chromium.org/svn/trunk/src@65095 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 669f5e7c1c3cb7c2343c0d3067b50d5fd0a10fc5