testserver.py listens on ephemeral ports by default.
If --port is specified on the command line, testserver.py will listen on that port, otherwise it will listen on an ephemeral port. If --startup_pipe is specified, the port number is written to the pipe as a 2 byte unsigned int in host order.

TestServer spawns testserver.py to listen on an ephemeral port and reads the port value from the pipe. A fixed port can not be specified.

BUG=56814
TEST=try bots pass

Review URL: http://codereview.chromium.org/4136008

git-svn-id: http://src.chromium.org/svn/trunk/src@65465 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 10e06c8a493378189b02bb3b0dc0ae43d0720d11