testserver.py listens on ephemeral ports by default.
If --port is specified on the command line, testserver.py will listen on that port, otherwise it will listen on an ephemeral port. If --startup_pipe is specified, the port number is written to the pipe as a 2 byte unsigned int in host order.

TestServer spawns testserver.py to listen on an ephemeral port and reads the port value from the pipe. A fixed port can not be specified.

BUG=56814
TEST=try bots pass

Review URL: http://codereview.chromium.org/4136008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@65465 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 69989047a3fdda3e8554a0c38ea4f8348e9094bb