Close URL requests associated with an instance when that instance goes away.
Not doing this causes a bug in shutdown (wehn we can't do fast shutdown and just kill the renderer).

The problem is that if a URLLoader is open (like maybe it's leaked by the plugin), the ResourceTracker will be the one to delete it. The ResourceTracker is a singleton which is deleted in the AtExit manager. If a URLLoader is deleted from there we'll crash trying to delete the corresponding WebKit loader which no longer exists (because the page associated with it went away).

BUG=61795
TEST=none
Review URL: http://codereview.chromium.org/4690003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@65501 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: bcaf27034e676598d8c8258f15f52f1783f265d3