net: assorted changes needed for Snap Start tests.
In IsAllowedBadCert, the pointer compare doesn't always work. Creating a
certificate by X509Certificate::CreateFromBytes and putting it in the
allowed bad list doesn't match the same certificate from a server.

InvalidateSessionIfBadCertificate: I'm sure that there's history here
that I don't know, so please review carefully. In order to test
resumption against a server with a test certificate we need to avoid
destroying the session.

ClearSessionCache, AdoptSocket: only for testing

BUG=none
TEST=none

http://codereview.chromium.org/4558004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@65713 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 05992906961618fe434543e84a868ca154f408a5