net: assorted changes needed for Snap Start tests.
In IsAllowedBadCert, the pointer compare doesn't always work. Creating a
certificate by X509Certificate::CreateFromBytes and putting it in the
allowed bad list doesn't match the same certificate from a server.

InvalidateSessionIfBadCertificate: I'm sure that there's history here
that I don't know, so please review carefully. In order to test
resumption against a server with a test certificate we need to avoid
destroying the session.

ClearSessionCache, AdoptSocket: only for testing

BUG=none
TEST=none

http://codereview.chromium.org/4558004

git-svn-id: http://src.chromium.org/svn/trunk/src@65713 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 072bef9f54209542a0de8607c8a7d776f9084c99