net: restart on NPN misprediction and add tests for Snap Start errors.
Restart an HTTP load when we mispredict the NPN protocol in a Snap Start connection.

Also, test the behaviour of the HTTP layer when SSL errors result
from the first Write() call. When we are doing Snap Start, we need to
merge the first application data record into the ClientHello. Thus the
SSL socket will report that it has connected immedately in order to get
a Write() call. This means that some errors, which would typically
result from the Connect(), are now returned on the first Write().

BUG=none
TEST=net_unittests

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@65837 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 7a342c221f326c3830cd37b5f8b2a36e499fb07e