Fix python sync server on Chrome OS
Bug 58522 was logged because bringing up the python sync server on Mac
failed due to a mismatch in the path to the protobuf files. We recently
noticed the same issue on our Chrome OS autotest runs as well.

This patch modifies the workaround checked in with
http://src.chromium.org/viewvc/chrome?view=rev&revision=62335 so that it
applies to Chrome OS as well.

BUG=60637, 58522
TEST=sync_integration_tests on Chrome OS

Review URL: http://codereview.chromium.org/4133001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@65880 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 412b2a3ddcc0c7536c1c8254645e8b18fcc73ad4