NSS: always record the ServerHello data.
Previously, if we predicted a ServerHello (because we were snap
startting), then serverHelloPredictionData was set and we wouldn't
record the actual ServerHello contents in the event of a misprediction.

This meant that we couldn't recover from a misprediction because we
would record the same prediction again.

With this patch, we'll always write to serverHelloPredictionData, even
if we made a prediction.

BUG=none
TEST=snap_start_unittests

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@65947 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 95d7c347e54bf812d3fb485d68bb081f53972d6f