[Chrome OS] Move dependency on cros.git into main DEPS file
Instead of depending on cros_deps (via a special .gclient rule) to
pull in a special DEPS file that expresses the CrOS dependency on
cros.git, just put that dependency in the main DEPS file.  We had
initially avoided this because we didn't want to force all the Chrome
devs on Linux to need to install git; now it seems like most of the
devs that work on Chrome Linux use git already.

BUG=chromium-os:8859
TEST=do a clean sync of the source tree, after taking the cros_deps clause out of your .gclient file.

Review URL: http://codereview.chromium.org/4876001

git-svn-id: http://src.chromium.org/svn/trunk/src@66126 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 5a531ad112aeded88a0acc06f0ec96232106100f