Added glResizeCHRM to resize the drawable associated with the command buffer.
This new command is used to pipe resize requests to the GPU process. It will allow is to synchronize the resize with the drawing. Specifically, if we resize at the wrong time, we will damage the backbuffer, causing visual artefacts (http://code.google.com/p/chromium/issues/detail?id=54430).

The resize request gets sent to the GLContext. Each backend will implement it differently. At this point it's set as a NOP.
-I will do an XResizeWindow on Linux.
-Nat will do a resize on Windows.
-Nico (or me?) will refactor the MacOSX code to use this path.

There is one line webkit patch that's necessary to activate this code (removing an #ifdef).

BUG=none
TEST=none
Review URL: http://codereview.chromium.org/4671003

git-svn-id: http://src.chromium.org/svn/trunk/src@66307 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: a623ff74ed14d43ef1179c2790d0ff306dfdbf15