chromeos: Fix screen locker tests.
These appear to have been following the pattern of:

  ScreenLocker::Show();
  ...
  ui_test_utils::WaitForNotification(
      NotificationType::SCREEN_LOCK_STATE_CHANGED);

I believe that this hangs until the test times out if the
notification is sent before we return from Show() (since we
won't yet be listening for the notification at the time).

I'm changing them to:

  ScreenLocker::Show();
  ...
  if (!chromeos::ScreenLocker::GetTester()->IsLocked())
    ui_test_utils::WaitForNotification(
        NotificationType::SCREEN_LOCK_STATE_CHANGED);

The old version looks like it started failing with my
r66287, which I don't understand -- I thought that we were
already generating the notification immediately as of
r66131, which fixed a bug where the initial attempt to grab
the pointer and keyboard always failed (thus ensuring that
the notifications wouldn't be sent until the next time that
the event loop was run).

BUG=none
TEST=all screen locker tests passed on my local machine

Review URL: http://codereview.chromium.org/5054004

git-svn-id: http://src.chromium.org/svn/trunk/src@66340 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 2b4f2552fa97511602135874f8677bb7d7636481