Fix omnibox update bug.
When tab-switching happens, the state of the tab is stored so that it can be restored later. However, when a background tab is being closed, it doesn't cause a tab-switch, and trying to save the state causes problems in some cases (e.g. 61588). So do not save state of a background tab.

BUG=61588
TEST=see bug

Review URL: http://codereview.chromium.org/5075001

git-svn-id: http://src.chromium.org/svn/trunk/src@66445 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: a929bc08d18bb9c5c9e39a0ee4fad5b6f1575d40