Fix SPDY crash on race when canceling a stream that just got created.
When I fixed the code not to be re-entrant (since that caused crashes) in r61880, I created a window when the pending create callback was posted to the MessageLoop to be run on the next iteration.  In this window before it actually gets invoked, if the pending stream creation got cancelled, then the callback wasn't cancelled, so we would execute a callback on a cancelled stream creation, which can cause crashes.

The fix is to keep track of the pending callbacks.  Cancellation of pending stream creations check this pending callback map first.

BUG=63532
TEST=See bug thread for repro steps.  New unit test added as well.

Review URL: http://codereview.chromium.org/5174005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@66630 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 79a7213e32d218505aa1367b80a01df541f86eb1