Fix a problem that would only show up in non-admin mode.
We were letting the result of IEIsProtectedModeURL decide the code flow of our tests, and for some reason, when run in non-admin mode, it causes a CoCreateInstance to be called and we mock it.

So we decided to mock IEIsProtectedModeURL too, so that we can have a more deterministic code flow, and now we have the opportunity (behind a TODO for now) to valid all code flows based on diferent return values.

TEST=Make sure the unit tests succeed in both and non-admin mode, which ever version of IE instance on whatever version of the OS.
BUG= 62387
Review URL: http://codereview.chromium.org/5116004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@66655 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: f7e92216ba19e490d213f6fc8b6ba3304a88861d