Fix a problem that would only show up in non-admin mode.
We were letting the result of IEIsProtectedModeURL decide the code flow of our tests, and for some reason, when run in non-admin mode, it causes a CoCreateInstance to be called and we mock it.

So we decided to mock IEIsProtectedModeURL too, so that we can have a more deterministic code flow, and now we have the opportunity (behind a TODO for now) to valid all code flows based on diferent return values.

TEST=Make sure the unit tests succeed in both and non-admin mode, which ever version of IE instance on whatever version of the OS.
BUG= 62387
Review URL: http://codereview.chromium.org/5116004

git-svn-id: http://src.chromium.org/svn/trunk/src@66655 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 7e82f955d2ca89fe23082bf13494b8ec8cff8ebd