Adding support for optional description in to the callback of the management API
Adding a list of API Permissions with the name "permissions" and adding a seperate list of Host permissions (i.e, http://*/*) so that developers can navigate them seperatly with ease. 

The permissions and hosttPermissions attributes are always visible in this object even if not supplied. Likewise the description attribute is also now "" if not defined in the manifest. As well as the options_url. 

Contributed by paulkinlan@chromium.org 
Original review: http://codereview.chromium.org/4585001

BUG=58388
TEST=chrome/test/data/extensions/api_test/management/*

Review URL: http://codereview.chromium.org/5211003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@66701 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 454203b6f65cbd5e730f2474d2a668edb3606aa8