Adding support for optional description in to the callback of the management API
Adding a list of API Permissions with the name "permissions" and adding a seperate list of Host permissions (i.e, http://*/*) so that developers can navigate them seperatly with ease. 

The permissions and hosttPermissions attributes are always visible in this object even if not supplied. Likewise the description attribute is also now "" if not defined in the manifest. As well as the options_url. 

Contributed by paulkinlan@chromium.org 
Original review: http://codereview.chromium.org/4585001

BUG=58388
TEST=chrome/test/data/extensions/api_test/management/*

Review URL: http://codereview.chromium.org/5211003

git-svn-id: http://src.chromium.org/svn/trunk/src@66701 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 17fe7b1e7856dd9914e3beb3e07ca86afc613418