Cleanup histogram synchronizer
Some of the assertions in histogram synchronizer, specifying
thread-to-run-on, were breaking. We used to use a
fancy mix of threads and locks to access the data
involving which sequence was pending, and this
transitions to clean use of locking only for
all members of the class.

BUG=62810
r=wtc
Review URL: http://codereview.chromium.org/5189002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@66861 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: fd11f04f344d99f47742b9267d6bcbebb5c8c81e