Make DeleteKey support GetLastError.
When SHDeleteKey returns, the last error value hasn't been set.
With this change, it's possible to have checks like:

if (!key.DeleteKey(...) && GetLastError() != ERROR_NOT_FOUND) {
  HandleErrorCase();
  return false;
}

return true;

As it stands, the last error value will always be ERROR_SUCCESS, which is not all that useful.

In fact there is code in the installer that attempts to do this, but doesn't currently work correctly (also because it's checking for the wrong GLE value, but I'm fixing that in a separate change).

BUG=none
TEST=Should have no side effects.

Review URL: http://codereview.chromium.org/5182009

git-svn-id: http://src.chromium.org/svn/trunk/src@67049 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 274a5eddfa2b4091c60539e7e23e5ad91ba06ebc