touchui: Create touch-events.
Create and process touch-events. The touch-events are created from
pointer-events. The way to decide whether a 'pointer event' originated from a
touch-device is not yet well-defined. So for now, use the --touch-devices
command line parameter to specify which pointer devices should be treated as
touch-device. For example, you can run:

  ./out/Debug/chrome --touch-devices=7,8

to treat the events coming in from devices with id 7 and 8 as touch events.
(these are the slave ids you get from 'xinput list')

BUG=None
TEST=None

Review URL: http://codereview.chromium.org/4738001

git-svn-id: http://src.chromium.org/svn/trunk/src@67133 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 7a3ded1d2f52442f0303700c993689d7f16420a7