New Cloud Print Proxy setup UI.
New UI to setup the proxy in Chrome.  This UI passed review.  New features
include a friendlier layout, a Learn More link, and a Print Test Page button on
the success page.  This does not yet bring the UI out from behind a flag; that
will be a separate CL.

BUG=60092
TEST=CloudPrintURLTest.CheckDefaultURLs

Review URL: http://codereview.chromium.org/5285001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@67602 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: a3fb7eebeea6f81b1d334f12b2b1b0e83e0d60ed