Add a base class for objects that want to filter messages on the IO thread.  I'll switch the filters to it in future separate changes.
I've also taken out the special case for an initial filter from the IPC classes.  The reason it existed was that there was a race condition of some messages not being filtered if a filter is added after construction but before launching the peer process.  Taking it out allows us to add more than one filter and makes things a little cleaner.
Review URL: http://codereview.chromium.org/5513001

git-svn-id: http://src.chromium.org/svn/trunk/src@68043 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: cfc49e7f4498395e6cb3accd75dcce3791eaa6bf