startup: move sandbox initialization together
The initialization call was happening on one place on Windows, then later on OS X.
This moves Windows down alongside OS X.

The comment above this code, dating back to 2008, says that we must initialize
the sandbox before the crash handler.  But this code as it currently works
runs way after we initialize breakpad (we do breakpad in WinMain, which
then loads this code from chrome.dll and runs it) and there is no breakpad
code in between its old location and its new one.

Review URL: http://codereview.chromium.org/5439003

git-svn-id: http://src.chromium.org/svn/trunk/src@68093 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 83bde11893fe8ff0c2c61af0539f05479f832e0e