Removed kChromiumRendererIdProperty window property.
This property was used by the GPU process to determine whether a particular renderer process had the right to render to a window. It turns out that this is no longer needed because now the browser tells the GPU process which child compositor window to render to and the GPU process trusts the browser process.

Also uncovered and fixed a bug in gpu_channel_host.cc where the renderer would hang if the GPU process failed to return a command buffer.

TEST=try
BUG=64834

Review URL: http://codereview.chromium.org/5607001

git-svn-id: http://src.chromium.org/svn/trunk/src@68096 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 8cbd0ff8edc6714f5db2c93acf53033212cb6ebd