Handle shift-selection in GroupTables.
Change the shift-click support in TableView so shift-clicks that span multiple
groups are prohibited. Clicking in one group, then shift-clicking in another
group will cause the shift-click to be treated as a normal click (i.e. just
select the item in the second group).

BUG=63787
TEST=Open task manager, click on item in one group, shift click on item in another group.

Review URL: http://codereview.chromium.org/5504001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@68123 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 42e1807b4ba4feb66bcd9b934720cf6610739260