chromeos: Refactor volume bubble for other settings.
This adds a new SettingLevelBubble class (nothing new here;
it's just the previous implementation of VolumeBubble but
with a configurable icon) and makes VolumeBubble implement
it.  I'll add a BrightnessBubble class in a later changelist,
along with code to hide the other bubble when we start
showing one.

BUG=chromium-os:4816
TEST=built it and checked that the volume bubble behaves the same as before when pressing F8, F9, and F10

Review URL: http://codereview.chromium.org/3766001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@68195 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: b60d724910978aa9216aefdf614cdc01c4e95dd9