Refreshing a ChromeFrame page which was received in response to a top level POST to a page in IE should reissue the POST request after bringing up a confirmation dialog.
This CL adds support to ChromeFrame to achieve this. Ideally we would want IE to display
the confirmation dialog for the POST. However the doc object host which is implemented by
IEFrame expects the doc object to implement the IHTMLDocument interface and a bunch of private
interfaces. 

To reissue the POST request we save away the POST data and headers in the navigation manager
instance which is per BHO. When we refresh the page in ChromeFrame the active document on
seeing that we have posted data issues a navigation back to the URL with the posted data
and saved headers. All other refreshes continue to work the same way as before, i.e. the
refresh request is sent to Chrome on the automation channel. The confirmation dialog is
put up by the active document with the same text and caption as that in Chrome.

Fixes bug http://code.google.com/p/chromium/issues/detail?id=64901

BUG=64901
TEST=Covered by new ChromeFrame unit test

Review URL: http://codereview.chromium.org/5595002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@68213 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 0baaaddad2d9e7881a08afc221448b02ccc8c9da