Allow painting multiple paint rects.
Previously, if there was no scrolling, then we would always paint the union of
any invalidations.  With this change, if the area of the invalidation rects is
sufficiently less than the area of the union, then we will just paint the
individual invalidation rects.

This may regress page cycler performance (see http://crbug.com/29589).

R=brettw
BUG=50303
TEST=updated unit tests

Review URL: http://codereview.chromium.org/5613002

git-svn-id: http://src.chromium.org/svn/trunk/src@68403 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 827694cbb7bbed514de802187c8e4c80f5fb7f65