chromeos: Show onscreen bubble when brightness changes.
This adds a BrightnessLibrary class that registers to receive
notification of brightness changes from the libcros code
that I added in http://codereview.chromium.org/5640003/,
a BrightnessBubble class to control an onscreen indicator of
the current brightness level, and a BrightnessObserver class
that wires up the two.

This change depends on
http://codereview.chromium.org/5603012/, which pulls the
updated version of libcros into Chrome.

BUG=chromium-os:8473,chromium-os:4816
TEST=built and tried it

Review URL: http://codereview.chromium.org/5620004

git-svn-id: http://src.chromium.org/svn/trunk/src@68879 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 460069927ba6be52feef2c1aae01f678f9179342