touchui: Fix painting issues caused by the X message pump.
WidgetGtk has two GtkWidgets, |widget_|, and |window_contents_| inside it.
|window_contents_| is the widget that receives the GTK+ expose events and shows
up on top of |widget_|. However, RenderWidgetHostViewViews paints on |widget_|s
window, because |widget_| is considered as the NativeView. So change
native_view() to return |window_contents_| instead of |widget_| so that the
correct canvas is used for painting.

BUG=None
TEST=Webpages show up correctly with the X message pump.

Review URL: http://codereview.chromium.org/5776002

git-svn-id: http://src.chromium.org/svn/trunk/src@69036 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 6f568941538c219b639e979ceb6d4e02e2be81a9