Fix pyauto downloads tests to trigger dangerous downloads to match expectation
It used to be possible to navigate to a dnagerous file type to trigger its
download. Not anymore. Now all downloads where the user initiated it (by
navigating to a URL or clicking on a link) are not marked unsafe even for
unsafe filetypes.

Update the tests to trigger dangerous downloads by navigating to a file which
in turn triggers the dangerous files to download.

BUG=66081, 57875

Review URL: http://codereview.chromium.org/5781006

git-svn-id: http://src.chromium.org/svn/trunk/src@69093 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: b470c40b058c7033ac7a1523bb3cd58529ce4845