render_view_browsertest.cc: Cast wchar_t to int for snprintf.
To print the character code, we should cast the wchar_t to an int.
The code currently happens to work because we don't compile with -fshort-wchars,
and thus wchar_t is four bytes, which coincides with what %d in the
snprintf format string expects. However, it's not portable, and Clang
complains about this.

BUG=none
TEST=browser_tests --gtest_filter=RenderViewTest.OnHandleKeyboardEvent

Review URL: http://codereview.chromium.org/5755005

git-svn-id: http://src.chromium.org/svn/trunk/src@69145 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: ecbd299f2a4f9f2575f7ac869a0bf1cb2d02548a