net: fix verification merge savings calculations.
Previously we miscalculated the amount of time saved by merging
certificate verification. We should have been counting the time from the
start of the verification till the time when we needed it because the
verification may have completed.

In the case that the verification has already completed, only count the
amount of time that it took.

BUG=none
TEST=none

http://codereview.chromium.org/5777005

git-svn-id: http://src.chromium.org/svn/trunk/src@69265 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 659c7d2f3785bc5f0b1afce5a9f4352599eeb3fc