net: fix verification merge savings calculations.
Previously we miscalculated the amount of time saved by merging
certificate verification. We should have been counting the time from the
start of the verification till the time when we needed it because the
verification may have completed.

In the case that the verification has already completed, only count the
amount of time that it took.

BUG=none
TEST=none

http://codereview.chromium.org/5777005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@69265 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 3c1d9a2c9dfdb3ac9c4fb390da2251eca7e579e6