Add support for temporarily trusting a certificate for the duration of unit tests on Windows, rather than requiring the machine to be pre-configured out-of-band.
Given the lack of a Microsoft-provided high-level API to supply application-level trusts to the verification routines, this implements a workaround that intercepts attempts to open the trusted system root store and injects the test certificates directly. This allows the unit tests to work without requiring that the Test CA be added to the machine's Trusted Certificates store.

While doing so, clean up the interface to adding/removing trusted test certificates, so as to support more than one trusted certificate if necessary.

BUG=8470
TEST=To follow

Review URL: http://codereview.chromium.org/4646001

git-svn-id: http://src.chromium.org/svn/trunk/src@69351 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 488a9df412ed1992aadbc7fa828c121059eb4a22