Revert 69428 - Add support for --multi-install to mini_installer, fixed a memory leak, fixed use of Win32 registry API, and refactored SetFullInstallerFlag so that it can be exercised in unit tests (tests to follow).
When --multi-install is not on command-line, "-full" is still added to the appropriate product "ap" value in official builds.  When --multi-install is present, "-full" is added to the multi installer's "ap" value unless a product is being migrated from single-install to multi-install, in which case it is put on the product's "ap" value as before.

BUG=61609
TEST=Manual for now: run mini_installer.exe in various configurations, inspecting "ap" values as you go.

Review URL: http://codereview.chromium.org/5875003

TBR=grt@chromium.org


This broke the Win Builder (dbg) (Shared) bot but no-one noticed.


Review URL: http://codereview.chromium.org/5890004

git-svn-id: http://src.chromium.org/svn/trunk/src@69462 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 35fedfec61ad8093ffe56b2f68f9cbd327b5d3f7