Another attempt at landing: http://codereview.chromium.org/5730003/ Previously r69187
Check-in was almost clean, but checkdeps.py flagged it as having invalid includes (because the Clang plugins aren't meant to be built as part of Chrome).  This adds the magic incantation (a DEPS file...  who knew?) to make checkdeps.py ignore it.  Tested by running checkdeps.py in my gclient before and after.

BUG=61004,62983
TEST=compile asserts and ppapi/tests/test_struct_sizes.c

Review URL: http://codereview.chromium.org/5927004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@69483 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 22e8abc49990e1d1c264d306a4b73ae4332c7e5e