Fix a crash when enabling hidden browser actions in incognito and then selecting Show button from the extension management page.
What was happening was that the extension was erroneously getting added to the extension toolbar model when the extension was reloaded the way enabling in incognito does. This caused the icon to show up in the toolbar and when you then explicitly show the extension it would be added again to the model. This leads to a crash when you uninstall (because one entry is removed and the other becomes dangling).

BUG=67201
TEST=None

Review URL: http://codereview.chromium.org/6075001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@69807 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 34bfcc0c870dd7b4c21a0d1b9ec757f5a3ddc9ee