touch: Fix switching focus among fields in webpages.
If Tab is pressed, then the RootView needs to process it first,
because the focus-manager will move the focus to the next focusable
view, without letting the currently focused view process it (which
means, for example, tab-ing to move focus between fields/links in a
RenderWidgetHostViewViews won't work). For all other keys, let the
focus manager process it first so that the keyboard accelerators can
be triggered.

BUG=pressing tab in a textfield in a touch-build doesn't move focus out of it.
TEST=see bug

Review URL: http://codereview.chromium.org/6016002

git-svn-id: http://src.chromium.org/svn/trunk/src@69846 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 75be6b9517b9a0ef43807be2f9c3382632945c5f