touch: Fix switching focus among fields in webpages.
If Tab is pressed, then the RootView needs to process it first,
because the focus-manager will move the focus to the next focusable
view, without letting the currently focused view process it (which
means, for example, tab-ing to move focus between fields/links in a
RenderWidgetHostViewViews won't work). For all other keys, let the
focus manager process it first so that the keyboard accelerators can
be triggered.

BUG=pressing tab in a textfield in a touch-build doesn't move focus out of it.
TEST=see bug

Review URL: http://codereview.chromium.org/6016002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@69846 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 1e86028b844186770e2850d579037e67653e423f