Adding proper authentication to the DOM based login screen.
This CL contains all of the additional code that is required to add proper
authentication to the DOM login screen. Specifically is adds a second sub-class
of AuthenticatorFacade called AuthenticatorFacadeCros that handles
authentication via libcros. Since the authentication is dependent on libcros,
this class is only used in chromeos builds, with the stubbed authenticator class
being used on non-chromeos builds. All of this code continues to be only
available in touchui builds. There have been additional tests added for the new
class and support in the exising tests to support both cros and non-cros builds.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/5708002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@69849 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: f5900c9ad70836519462f3bae5e140ac50c3fe2f