Prevent App extensions to be returned by AutomationProvider::GetEnabledExtensions().
These are not useful for testing and are causing CEEE problems, since it expects a single extension being installed.

A better fix might be to not have any default extension installed with Chromium and only have them with a branded Chrome so that they don't show up in Chrome Frame or CEEE installs... Maybe... Started a conversation with the Chrome Apps and Extensions team to discuss that.

BUG=None
TEST=None

Review URL: http://codereview.chromium.org/6008004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@69867 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 8b3c758702bd9c82bf1f6639dcbbf1bf56e13c7d