Revert 68944 - Revert "Add named testing interface."
Reverting the revert of named interface CL. Followup CL which addresses Pawel's comments is up at
http://codereview.chromium.org/5967003/

http://src.chromium.org/viewvc/chrome?view=rev&revision=67300

The change has been approved conditionally,
i.e. if the outstanding review comments are handled
in a follow-up as the top priority.

Two weeks have passed, and I have not seen the follow-up.
It is risky to keep this change in the tree, as it is already
non-trivial to revert.

Conflicts:

  chrome/browser/automation/automation_provider.cc
  chrome/test/ui/named_interface_uitest.cc
  ipc/ipc_channel_posix.cc

TBR=nirnimesh

BUG=chromium-os:8512
TEST=none

Review URL: http://codereview.chromium.org/5707006

TBR=phajdan.jr@chromium.org

BUG=chromium-os:8512
TEST=ui_tests --gtest_filter=NamedInterfaceTest.FLAKY_BasicNamedInterface

Review URL: http://codereview.chromium.org/6012006

git-svn-id: http://src.chromium.org/svn/trunk/src@70089 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: a35263b1e4f28bf51f1d7dc9c1ff0c70d90cc09c