Do machine inspection on install.
When installing Chrome [x]or Chrome Frame in multi mode, check if the other product (CF xor Chrome) is installed.
If it is, also include it in the installation as we need to update both products in sync.

For Chrome Frame, I also added checks to see if CEEE is already enabled or if the user has already opted into GCF.
If CEEE has been enabled, we implicitly include it when the install is run even though ceee is not on the command line.
If ready-mode is set on the command line but Chrome Frame is already installed without it, we won't enable ready mode.

TEST=Try installing Chrome Frame in multi mode.  Then install a more recent version of Chrome in multi mode.  Both product should be updated.  Secondly, for Chrome Frame, install it first and then try to run the installer again, but with ready-mode ont the command line.  Chrome Frame should not be switched to ready mode.
BUG=61609

Review URL: http://codereview.chromium.org/6091008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@70268 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: cd72d9f07702429fb47d750b8d364b5a24ede0e4