views: Improve cursor movements on word boundaries.
(1) When there's blank space at the end of the string in the textfield and the
cursor is at the end, 'move cursor to previous word' should move the cursor at
the beginning of the last word.
(2) When there's blank space at the end of the string, and the cursor is at the
end of the last word, 'move cursor to next word' should move the cursor at the
end.
(3) Moving the cursor should make the cursor immediately visible. This is more
inline with how GTK+ cursors behave. For example, constantly pressing the
home-key in a GTK+ textfield makes the cursor visible for the duration of the
keypress.

BUG=none
TEST=NativeTextfieldViewsTest.CursorMovement

Review URL: http://codereview.chromium.org/5972008

git-svn-id: http://src.chromium.org/svn/trunk/src@70468 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 3e93eb4dcd382bc28551de821cd838651272b766