Enable "inspect element" for chrome:// pages.
It's already possible to open the js debugger for these pages via the wrench menu, so disabling "inspect element" is just a nuisance. Also, there's no obvious reason in my mind why it should be different for debug/release (I do most of my domui development in release as it's faster to build and easier to use the debugger).

BUG=none
TEST=manual

Review URL: http://codereview.chromium.org/6094007

git-svn-id: http://src.chromium.org/svn/trunk/src@70540 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 1f59020a4c03d37b4b25d0c195fcb231db24259c