Enable "inspect element" for chrome:// pages.
It's already possible to open the js debugger for these pages via the wrench menu, so disabling "inspect element" is just a nuisance. Also, there's no obvious reason in my mind why it should be different for debug/release (I do most of my domui development in release as it's faster to build and easier to use the debugger).

BUG=none
TEST=manual

Review URL: http://codereview.chromium.org/6094007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@70540 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: fdc429ef30a1f4b6d601b6861a9d6f5213379285