roll skia 636:673
I need r666 (r667 and r668 are build fixes for r666. r673 fixes a regression that was introduced in r637 – found by the chromium trybots).

Some code changes required due to r637, which changed the signature of a pure virtual method and the signatures of several non-virtual methods. I added OVERRIDEs on windows to let this be a compile error in the future.

I also added SkBitmapCache.h/cpp to the gyp file which was added in r655. SkGradientShader depends on this.

BUG=none
TEST=none

Committed: http://src.chromium.org/viewvc/chrome?view=rev&revision=70693

Review URL: http://codereview.chromium.org/6081006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@70835 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: cbfa2b89779dcb121a54485a1c0909868e9dbd5c