Revert 70967 - Full refactor and rewrite of the autofill sync integration tests
Reason for revert: Need to merge with patches that slipped in before this.

The autofill sync integration tests were written more than a year ago,
and are pretty badly out of date. In addition, the autofill test
framework is crufty and adding new tests is not intuitive.

The autofill sync code has recently undergone a major refactor, and so
it is time to update the integration test code too. This patch contains
a major update of the LiveAutofillSyncTest class and a rewrite of most
of the autofill sync tests.

BUG=62913, 60368
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/6163001

TBR=rsimha@chromium.org
Review URL: http://codereview.chromium.org/6166006

git-svn-id: http://src.chromium.org/svn/trunk/src@70969 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 0689de81892e37a4e74cbbda4c7bffe3f2085f6b