Re-land: Full refactor and rewrite of autofill sync integration tests
The autofill sync integration tests were written more than a year ago,
and are pretty badly out of date and inaccurate. In addition, the
autofill test framework is crufty and adding new tests is not intuitive.

The autofill sync code has recently undergone a major refactor, and so
it is time to update the integration test code too. This patch contains
a major update of the LiveAutofillSyncTest class and a rewrite of most
of the autofill sync tests.

NOTE: This was originally code reviewed in http://codereview.chromium.org/6163001. This is an attempt to re-land the patch after merging with existing code in the tree.

BUG=62913, 60368
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/6130005

git-svn-id: http://src.chromium.org/svn/trunk/src@70973 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 62de06a42aab4a80c7acea44f9b2fe5caa4fbbe0