Updating expectation of autofill sync tests when a null profile is added
In the past, it used to be possible to add an autofill profile that had its
label field populated with all other fields blank.

Today, if a sync test were to add an autofill profile with only a GUID
(leaving all other fields blank), the operation is expected to be unsuccessful.

This patch updates the sync test framework with this new expectation.

TBR=lipalani@chromium.org
BUG=69170
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/6230001

git-svn-id: http://src.chromium.org/svn/trunk/src@70995 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 0581616b76a3e446a0587829374504758a3dd99f