net: Ensure that when using False Start + client auth, bad client certificates are not cached.
If an SSL handshake fails when client certificates are used, ensure that
the client certificate selected is removed from the SSL client auth
cache. This ensures that the user is prompted to select a certificate
again, as the cause of the failure may have been due to selecting the
wrong certificate or selecting no certificate when one is required.

The existing logic worked when TLS False Start was disabled, but could
fail when False Start was used or when the peer requests renegotiation.
This changes ensures the client certificate is removed from the cache by
moving the cache removal layer from the HttpStreamRequest to the
HttpNetworkTransaction.

Patch by: Ryan Sleevi

BUG=66424
TEST=HttpNetworkTransactionTest.ClientAuthCertCache*

git-svn-id: http://src.chromium.org/svn/trunk/src@71071 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: f078e897fed1e7fa432c218c0df8cc604c9909f8