give tabbed options clear browsing dialog a facelift
It should now match the mocks modulo the adobe flash settings link (not sure why the mocks show it as a checkbox; AFAIK making it a checkbox is outside our ability to do).

also make it work (the time period select dropdown was broken).

BUG=69163,63843
TEST=manual

Review URL: http://codereview.chromium.org/6110006

git-svn-id: http://src.chromium.org/svn/trunk/src@71090 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 683935d1a0d48c2a952fd9352cdfb4804c5cc590