give tabbed options clear browsing dialog a facelift
It should now match the mocks modulo the adobe flash settings link (not sure why the mocks show it as a checkbox; AFAIK making it a checkbox is outside our ability to do).

also make it work (the time period select dropdown was broken).

BUG=69163,63843
TEST=manual

Review URL: http://codereview.chromium.org/6110006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@71090 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: eeb386346086fa85bf121f82f2a2a4658e13446e