[Chrome OS] Fix a couple of problems around CAPTCHA handling upon password changes
This CL fixes two problems with CAPTCHA handling in the case where we
are attempting to re-auth with a new password.  First, in cases where
repeated auth attempts triggered a CAPTCHA, we were not reporting this
to the UI correctly.  Second, in cases where the user was re-prompted
for her password AND was in the CAPTCHA state, we weren't keeping
track of the captcha token properly, and so we were not actually able
to clear the CAPTCHA at all.

BUG=chromium-os:10155
TEST=unit tests, install on device and try it

Review URL: http://codereview.chromium.org/6190004

git-svn-id: http://src.chromium.org/svn/trunk/src@71204 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 09700677db4c8009523094b2d9f0e62dd25522d7