Disable background network access in sync integration tests
Now that the sync integration tests are designed to be self-contained
and not rely on external network resources like GAIA, we need to ensure
that background network access for things like auto update is disabled.

This patch adds the --disable-background-networking command line switch
to the sync integration test framework.

BUG=69469, 53931
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/6282001

git-svn-id: http://src.chromium.org/svn/trunk/src@71338 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 934656d7fe34966698a031090ba89f5679467933