Debugging updates for sync integration tests
The sync tests were using verbose logging in the past in order to help
debug flaky and failing tests. Now that the tests are mostly green, the
verbosity can be dialed down. This patch reduces the verbosity of sync
test logging.

In addition, a couple of bookmark tests occasionally fail due to a
mismatch, and the logs don't make it clear where the test fails. This
patch adds more detailed logging during bookmark comparison.

TBR=tim@chromium.org
BUG=none
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/6351001

git-svn-id: http://src.chromium.org/svn/trunk/src@71456 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 29a59c56446b788d476c5383cba0342ed83c414c