tabbed options - new clear browsing data overlay
I tried to make this look as nice as possible (and close to the mocks), but Alex will have to fiddle with the CSS to finalize it. I didn't implement the button changing to "clearing..." because it changes the width of the button and thus the whole overlay. The throbber + disablement should hopefully be enough of a visual cue.

BUG=63843
TEST=manual

Review URL: http://codereview.chromium.org/6340002

git-svn-id: http://src.chromium.org/svn/trunk/src@71703 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: fd12b7adc6bf069411b54464b12a8a3c38dbde61