Cancel any pending speech recognitions when the dispatcher host terminates.
I have not added a new test for this because the existing test is flaky and
didn't seem right to add another flaky test like that. Instead I've added a TODO
to add a browser test for this case once the flaky test has been fixed.

BUG=none
TEST=manual

Review URL: http://codereview.chromium.org/6358007

git-svn-id: http://src.chromium.org/svn/trunk/src@71956 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: ad0d6c1b96d4713e99426d4ea0d153d7c244075a