Gpu process restart fix.
If the gpu process dies, CommandBufferProxy might be the first one to find out about it.  When the context is being created, GpuChannelHost still thinks there is no error.  It will send CreateViewCommandBuffer message to establish a communication channel, which will fail, and GPU process will not restart.  To prevent that, we set GpuChannelHost state to lost as soon as CommandBufferProxy finds out about it.

BUG=58785
TEST=open a composited tab, in another tab go to about:gpucrash

Review URL: http://codereview.chromium.org/6384003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@72023 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 69083419fdedb2a72b3efabcfed1a3fc245ac970