The function is now only used with the CHROME_OMNIBOX point and the value is cached and returned from the cache on subsequent calls. All other access points will be always freshly fetched and returned.
Future calls to this function with another access point should either be done on the FILE thread or implement caching the way it is done for Omnibox. Moreover if they implement the caching mechanism they should be prepared     for using this function in asynchronous manner if called on another thread.

BUG=62337
TEST=A search from branded browser through the omnibox-the rlz parameter should be appended to the query string.

Review URL: http://codereview.chromium.org/6028012

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@72337 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: f0a21806cc5bfd066dd8d926366e84bf0eb522b6