Fix for crash when deleting |DataSource|s on the IO thread.
Although |ChromeURLDataManager::DataSource|s are |RefCountedThreadSafe|, at least two derived classes contain |CancellationFlag| members, which check that they are deleted in the same thread as they are created.

Normally, the |DataSource|s are destructed when |ChromeURLDataManager| (instanced via a Singleton) is destroyed, on the main thread.  This is not an issue as the classes with a |CancellationFlag| are created on that thread too.

However, occasionally (see the bug) |AddDataSource()| is called when there is already an entry in |data_sources_|, leading to a |DataSource| being released and destructed in the IO thread.

To fix this problem, I've incremented the reference count and posted a task to the main (UI) thread that will release the reference and so destruct it there.

BUG=49121
TEST=none

Review URL: http://codereview.chromium.org/6285002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@72398 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 3a1699bf1c8571f27bb5bac7e35c2222cd46963f