Fix for crash when deleting |DataSource|s on the IO thread.
Although |ChromeURLDataManager::DataSource|s are |RefCountedThreadSafe|, at least two derived classes contain |CancellationFlag| members, which check that they are deleted in the same thread as they are created.

Normally, the |DataSource|s are destructed when |ChromeURLDataManager| (instanced via a Singleton) is destroyed, on the main thread.  This is not an issue as the classes with a |CancellationFlag| are created on that thread too.

However, occasionally (see the bug) |AddDataSource()| is called when there is already an entry in |data_sources_|, leading to a |DataSource| being released and destructed in the IO thread.

To fix this problem, I've incremented the reference count and posted a task to the main (UI) thread that will release the reference and so destruct it there.

BUG=49121
TEST=none

Review URL: http://codereview.chromium.org/6285002

git-svn-id: http://src.chromium.org/svn/trunk/src@72398 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 9ee08ddc9d003d856557986d089cb9707f7d354b