Make the sync integration tests self-contained on autotest
In the past, the sync integration tests used to require a password file
stored on every test device in order to do a gaia sign in using
production gaia servers. This caused the tests to be brittle.

As of today, the sync integration tests no longer rely on a password
file, with gaia sign in being stubbed out locally.

This patch reconfigures the tests on autotest, so that it no longer
looks for a local password file.

In addition, the tests run much faster now, and therefore, we reduce the
max timeout to a more reasonable 2 minutes (in the extreme case).

BUG=chromium-os:11294, chromium-os:9262
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/6387004

git-svn-id: http://src.chromium.org/svn/trunk/src@72561 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: e54b28430f7b301e04eb5b02ce667019df4434bf