Policy: generate boilerplate policy type and constant code.
This is the first step to being able to re-constitute policy from a non-generic protobuf in a policy request in a fully automated way (i.e. the code for the process can be generated). The non-generic protobuf is required for automatic merging of policy on the server-side.

BUG=68309
TEST=existing policy/provider tests

Review URL: http://codereview.chromium.org/6002015

git-svn-id: http://src.chromium.org/svn/trunk/src@72784 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 57af84f84f537a4e9748572921fda2f1d3a73952