Revert 72784 - Policy: generate boilerplate policy type and constant code.
This is the first step to being able to re-constitute policy from a non-generic protobuf in a policy request in a fully automated way (i.e. the code for the process can be generated). The non-generic protobuf is required for automatic merging of policy on the server-side.

BUG=68309
TEST=existing policy/provider tests

Review URL: http://codereview.chromium.org/6002015

TBR=danno@chromium.org

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@72785 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 07350d8b93656415b4dd0cc09df2abe98b476400