Unit tests that use ProfileSyncServiceMock are leaking due to adding a DataSource (SyncResourcesSource) to ChromeURLDataManager.
The basic problem is that the destructors need to be called on the UI thread, but at shutdown the UI thread (which is the only thread left) is not recognized so the destruction is thrown to the thread even though the message loop is no longer running.  This means that the messages are dropped and memory is leaked.

I added cleanup code to the destructor.

BUG=70782
TEST=Valgrind

Review URL: http://codereview.chromium.org/6317017

git-svn-id: http://src.chromium.org/svn/trunk/src@72876 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 7833858d7d1b64f9fbefcf461de72b5c22008f93