Make BookmarkModelVerifier wait for favicons to load before extracting them
The test case for bookmark favicon sync was failing in its verification
step because GetFavicon sometimes returned a blank favicon for a URL.
This was because GetFavicon can be asynchronous if the favicon of a
bookmark node is not already loaded.

This patch modifies the method that compares favicons, and takes care of
cases where the favicon has not yet loaded by adding a conditional wait.

BUG=69694
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/6347016

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@72877 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 59902cf5fd35518fb2cbeef3b708e8124015083e