Refactoring to use a new IPC message filter.
There were several ViewHostMsgs being sent to the GpuProcessHost via a RenderMessageFilter. This patch creates a GpuMessageFilter and renames the ViewHostMsgs to GpuHostMsgs. This will clean up some of the clutter in render_messages_internal and render_message_filter. It will also allow us to easily add new messages to the GpuProcessHost in the future.

BUG=none
TEST=by hand and trybots

Review URL: http://codereview.chromium.org/6401003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@73152 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: dc2e8859926f9954f55e4455d5bd5ec46a5abc70