Fix a possible access to a deleted MetricsLog object.
In the case where StopRecording is called and the log is over the event limit, the current_log_ is deleted and a new one is created.  However, the local variable current_log was still pointing to the old object.  Fixed by moving the initialization of current_log to after the log is recreated.  This was introduced in http://codereview.chromium.org/2731006 .

BUG=64813
TEST=none

Review URL: http://codereview.chromium.org/6347034

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@73251 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: fe4ba09dfea09c28d567ac1ea5945098c0ad3109