Silence testserver logs when tests are run in non-verbose mode
The console output from testserver.py, chromiumsync.py and xmppserver.py during sync integration test runs is pretty chatty, and could do with being disabled by default. In addition, testserver.py spits out logs for non-sync tests even when the tests are run in non-verbose mode.

This patch adds a switch to testserver.py called "log-to-console", and ties it up to the logging level returned by logging::GetMinLogLevel(). This way, testserver logs are printed on the console (in addition to being written to testserver.log) when tests are run in verbose mode. However, when tests are run in non-verbose mode, testserver logs are only written to the log file, but not to the console.

BUG=71241
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/6404003

git-svn-id: http://src.chromium.org/svn/trunk/src@73330 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: d4a5a29fd11f549da220d190ebae89cabb6a5c5e