Add presubmit checks for grit.  This runs the unittests.
Fix the grd_reader tests to pass if run from any directory (previously, it
had to be run from tools/grit).

Fix rc_unittest to pass on Linux (it was sensitive to line endings).

If the user doesn't have access to the internal repository, there will be
lots of failures :(  Maybe there's a way to detect this case and skip the
tests?

Review URL: http://codereview.chromium.org/6349026

git-svn-id: http://src.chromium.org/svn/trunk/src@73498 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: b9d8abb8c92968461ecf8a0f6c5d68ff4fd264fe