Refactoring GpuProcessHost for use by the browser.
This is a step towards allowing the browser to make requests of the GpuProcessHost (it's traditionally done by renderers). This is done with a level of indirection. It replaces the sending of IPCs to the renderer with calling a callback (which now sends the IPCs). The IPCs sent from callbacks originating in GpuMessageFilter use WeakPtrs to guard against the filter disappearing before the callback is called.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/6287026

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@73624 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: efde831020ef82fe1f6f5db972418c6c5d2dfda7