Makes it so the hung renderer dialog isn't shown if the page that is hung is being shown by instant.
This also fixes a crash if the instant preview hung. The crash occurred because when HungRendererDialogView::ShowForTabContents set the bounds a synchronous focus change was triggered, which caused the instant preview to be destroyed, and when  hung_pages_table_model_->InitForTabContents(contents) was invoked the contents was destroyed and no renderer.

BUG=58853
TEST=see bug

Review URL: http://codereview.chromium.org/6286055

git-svn-id: http://src.chromium.org/svn/trunk/src@73636 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: a04096a740c179639e7897d4b24d69b5824fb622