gtk: simplify menuing API
Bug 71821 looks like it was caused by using the wrong menu API, but
it wasn't obvious because there were so many entry points.  This
change simplifies it to two main entry points -- one for context
menus (which belong at the mouse position) and one for button-initiated
menus (which belong alongside the button, like the wrench menu).

This also makes the right-click menu in the task manager come up on
mouse-down on GTK, as that is how right-click menus should work.

BUG=71821
TEST=bring up a notification, click the wrench, pick an item from
the submenu; the selection should take effect

Review URL: http://codereview.chromium.org/6250145

git-svn-id: http://src.chromium.org/svn/trunk/src@73726 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 73f3c73130d6898601048b5189a04528af3cc9bd