GPU process does not fall back to alternative GL implementation.
Unless the --use-gl command line switch is specified, only the default GL implementation will be used. On windows this is ANGLE. On other platforms it it OpenGL.

Falling back on OpenGL implementations that are generally not as reliable as the default was causing crashes.

TEST=try, check GL implementation does not fallback
BUG=none

Review URL: http://codereview.chromium.org/6246116

git-svn-id: http://src.chromium.org/svn/trunk/src@73857 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: aab5dd22bea2c05b669527eda0143feadffebb90