When performing SSL client certificate authentication on Windows, and an operation using the certificate's private key fails, remove the cached provider handle from the client certificate.
If a user is using a smart-card based certificate, and removes the card from their reader, SSLClientSocketNSS will rightly fail the handshake because it's unable to generate the CertificateVerify message as part of the client certificate handshake. However, because the HCRYPTPROV used to obtain the private key is optimistically cached in the X509Certificate, all future attempts to use the private key, even after the card has been re-inserted, end up failing because the cached handle was invalidated once the card was removed. By removing the cached provider handle from the certificate, the user should be able to re-authenticate with the card/reader and, as a result, re-authenticate with the peer.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/6250042

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@73914 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 2463b9d24ce0bfbb7d70c2bac9c17d02170d6c26