NTP: Fix click issues with about: and file: most visited thumbnails
The reason these were not working correctly was that the handler for file: and
about: links was attached using capturing so the handler was always triggered.
This changes the events to use bubbling instead. This allows other listeners to
preventDefault before handleLinkClickOrMouseUp is reached.

BUG=72213
TEST=On the NTP
     Show most visited and visit some about: or file: URL enough for it to show up.
     Try to blacklist or pin the item.
     We should not navigate to the URL.

Review URL: http://codereview.chromium.org/6246146

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@74033 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: c3ab66f20248f455a422f83dc95936c3d41ef73f