Move most of GpuProcessHost processing to GpuProcessHostUIShim.
This is a refactoring to support the browser using the GPU process for drawing. Currently most of the communication with the GPU process (via GpuProcessHost) happens on the IO thread (where requests come in from the renderer). In the future, the browser will be drawing from the UI thread. Rather than have two entry points (or a lot of explicit thread hopping), I've moved GpuMessageFilter (the proxy for the renderer) to the UI thread and most of GpuProcessHost functionality to GpuProcessHostUIShim.

This cleans up some of the CVCBThreadHopping mess from an earlier commit.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/6349079

git-svn-id: http://src.chromium.org/svn/trunk/src@74192 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 759eda0aeb0e49defae7d139cee0a643829c5b53