Stop SystemKeyEventListener sooner
The worker thread created in AudioMixerAlsa should now stop and join as expected now that it is being stopped on the UI thread.  Previously, AudioMixerAlsa shutdown was done when Singletons are destroyed, during AtExitManager, which happens much later and off of the UI thread.

BUG=chromium-os:11745
TEST=Manual, Crash Report - Stack Signature: -192EB7B should stop occurring

Review URL: http://codereview.chromium.org/6465002

git-svn-id: http://src.chromium.org/svn/trunk/src@74212 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 398762720690ceffda58f1f6eea9bd1eb1eacfd4