Remove modality of MULTI_VALUE choice types in AboutFlags
Previously the UI for multivalues in AboutFlags was confusing, since
you could "enable" or "disable" the feature, then select an option
from a combobox.  This change removes this modality, and instead, the
first option from the multivalue should be considered the "disabled"
version of the lab (we even assert that it has no command line flags).

This CL was originally http://codereview.chromium.org/6410082/ , but
cbentzel suggested we split it up into two CLs: this is the first of
them.

BUG=none
TEST=AboutFlagsTest.MultiValues

Review URL: http://codereview.chromium.org/6413037

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@74277 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 76829d1d09ce45c04d909b5f0d157e937dd0d637