PhishingClassifierDelegate test cleanup, and fix a bug with session history navigation.
As of r74028, the RenderView will send notifications directly to the PhishingClassifierDelegate that we are using in the test.  As a result, we don't need to explicitly send DidCommitProvisionalLoad notifications to the delegate; the RenderView is doing that for us.

As a side-effect of replacing these one-off notifications with real navigations, I also noticed a problem where pending classifications are not cancelled on back/forward/reload because these are not new navigations.  This CL fixes that as well.

BUG=none
TEST=PhishingClassifierDelegateTest

Review URL: http://codereview.chromium.org/6425003

git-svn-id: http://src.chromium.org/svn/trunk/src@74345 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 762ab9a5b01987873bc4781f8da39d73124bdbfb