Get rid of BaseTabStrip::GetPreferredHeight
In an ongoing effort to simplify BaseTabStrip, I observed that GetPreferredHeight is an extra method of the BaseTabStrip base class that is only used to retrieve the height quantity of the GetPreferredSize.  GetPreferredSize is a View method, so I changed the callers to call GetPreferredSize().height() instead and removed this method.

In a subsequent CL, I plan to refactor so that the frames can interface to view::Views::GetPreferredSize instead of BaseTabStrip::GetPreferredHeight, thus generalizing the interface and reducing dependency on BaseTabStrip and its design.

This first step just cleans up BaseTabStrip.  The next CL will simplify the frame views.

BUG=None
TEST=None

Review URL: http://codereview.chromium.org/6312196
Patch from Chad Faragher <wyck@chromium.org>.

git-svn-id: http://src.chromium.org/svn/trunk/src@74356 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 7bb937d5c5109ff836b92342490a2cd1e42c50df