Fix flaky browser_tests incoming_queue_.empty() failure
AudioMixerPulse was sometimes being destroyed while the Init call was still in it's worker thread's queue.  If we see that Init hasn't been called yet, RunAllPending() is called to clear out the queue first.  This is goes along with the previous fix, where we are trying to close the AudioHandler sooner in the shutdown process  http://codereview.chromium.org/646500

BUG=chromium-os:11745
TEST=browser_tests should not fail with "check failed: incoming_queue_.empty()"

Review URL: http://codereview.chromium.org/6476004

git-svn-id: http://src.chromium.org/svn/trunk/src@74387 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 62d0d60bb4a04fbdd37665c0a9aa006b5260f11a