Add support for ICU data in a data file.
- Mac Foundation utility for fetching a resource from the MainAppBundle.
- Bringing in newer ICU dep that has optional support for using the ICU data
  file.
- If Mac is built using a ICU data file, directly map in the data so it will
  work even in a sandboxed process.

**The build does not use this data file support at this time.**

This might help the Mac bots as it means each unittest that depends on base
will be ~10meg smaller, helping link and also helping with the copy of bits
from the builder to the tester.

If Mac ever support a Universal binary (32bit and 64bit) it will also probably
want to use this to avoid carrying two copies of the ICU data.

Review URL: http://codereview.chromium.org/6479021

git-svn-id: http://src.chromium.org/svn/trunk/src@74474 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 839a207fc0bcf005e76642bb4210d43a68ed4d5b