Introduce OffTheRecordProfileIOData and ProfileImplIOData.
They both inherit from ProfileIOData. The former is for the off the record (incognito) profile. The latter is for the normal ProfileImpl profile.
All of the IO related Profile objects are now initialized at the same time, in the subtype implementations of ProfileIOData::LazyInitializeInternal().
I also took this opportunity to clean URLRequestContext up so it is a class and keeps its member variables private. This required touching a fair number of files.

TODO: Remove lots of the refcounting of member variables, since they can now be owned by ProfileIOData.

BUG=67237
TEST=none

Review URL: http://codereview.chromium.org/6286133

git-svn-id: http://src.chromium.org/svn/trunk/src@74561 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 0a9ca0ece0645f10f9185b94fc192fae34321de4