Revert 74583 - Revert 74571 - Use SSE2 block differ for chromoting
(Quick test to see if it is the cause of crashes in media_unittests)
(Test showed this CL is not to blame)

We have the SSE2 lying around in the tree just never being used. This will
allow us to use it.

BUG=None
TEST=Chromoting to a host machine and the diff will work correctly

Review URL: http://codereview.chromium.org/6469022

TBR=hclam@chromium.org
Review URL: http://codereview.chromium.org/6488023

TBR=finnur@chromium.org

git-svn-id: http://src.chromium.org/svn/trunk/src@74588 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 244151ae9d88cfbc49133788d121c8f76733ef13