clang: Make it possible to run Elliot's plugin on the bots
To make it possible to change the exact plugin command line without a master restart, the command line is loaded from a bash script. This script is used if clang_use_chrome_plugins and clang are both set in GYP_DEFINES when runhooks runs.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/6461026

git-svn-id: http://src.chromium.org/svn/trunk/src@74597 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 129a1b54df5f6cda2c6ed8a6365af81be1b459e9