Revert 74571 - Use SSE2 block differ for chromoting
We have the SSE2 lying around in the tree just never being used. This will
allow us to use it.

A number of Windows bots have gone red in media_tests on the waterfall:

http://build.chromium.org/p/chromium/builders/XP%20Tests%20%281%29
http://build.chromium.org/p/chromium/builders/XP%20Tests%20%281%29/builds/501/steps/media_unittests/logs/stdio

This seems to be related to r74571.  So am reverting to see.

BUG=None
TEST=Chromoting to a host machine and the diff will work correctly

Review URL: http://codereview.chromium.org/6469022

TBR=hclam@chromium.org
Review URL: http://codereview.chromium.org/6502002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@74630 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 22612fe801ff5ae98cf83f7267f72e1cb80869e8