Revert 74571 - Use SSE2 block differ for chromoting
We have the SSE2 lying around in the tree just never being used. This will
allow us to use it.

A number of Windows bots have gone red in media_tests on the waterfall:

http://build.chromium.org/p/chromium/builders/XP%20Tests%20%281%29
http://build.chromium.org/p/chromium/builders/XP%20Tests%20%281%29/builds/501/steps/media_unittests/logs/stdio

This seems to be related to r74571.  So am reverting to see.

BUG=None
TEST=Chromoting to a host machine and the diff will work correctly

Review URL: http://codereview.chromium.org/6469022

TBR=hclam@chromium.org
Review URL: http://codereview.chromium.org/6502002

git-svn-id: http://src.chromium.org/svn/trunk/src@74630 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: ba95e16a88eddb70fffb1647a78cdf8c2aa838e0