A number of poorly written IE BHO's crash IE if ChromeFrame is the currently loaded document. This is because they expect ChromeFrame to implement interfaces like IHTMLDocument2 on the same lines as regular IE documents. Currently in ChromeFrame we patch the invoke methods of these BHO's prior to firing navigation events from ChromeFrame. However this is not enough as these objects also crash for regular navigation events fired from IE when ChromeFrame is loaded.
We now don't fire navigation events for buggy BHO's if ChromeFrame is the current document.
The BuggyBho handler instance is now created once for the thread. We patch when we receive
navigation notifications from Chrome as before. When we receive a notification
on our patched event sink we check if CF is loaded and if yes skip the call. 

Added helpers to chrome frame utils to check if CF is loaded in the current web browser instance.

BUG=55932
TEST=none

Review URL: http://codereview.chromium.org/6493002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@74691 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 6b71834ecd324409290883d8b621ffdbbb7c28ec