Revised SSE2 color conversion
I discovered there's average and shuffle instructions which simplify the
calculations of UV components. Also the extract instruction was very useful.

In terms of the algorithm I rearranged the memory reads to be sequential.

Overall this gives about 10% improvement on 64-bits machine.

On 32-bits machine this gives 20% improvement. If compiled as 32-bits code this is
quite slow compared to 64-bits, about 2x slow down. So working on MMX code will
be a direction for improvement.

BUG=None
TEST=I tried it, it looks right.

Review URL: http://codereview.chromium.org/6488014

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@74827 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 0a68de24ccdb748c85f78245bfef9a886f922a2b