Add plumbing to webRequest API to support event filtering.
This introduces a new kind of event (WebRequestEvent) to handle the extra parameters to addListener. Behind the scenes, adding a listener creates a unique sub-event associated with that set of filter+extraInfo parameters. When we dispatch the event, we dispatch only those sub-events that match the required filters.

BUG=60101
TEST=no

Review URL: http://codereview.chromium.org/6250152

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@74887 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 918b251a641582c810141bf7f983ee2f1790be0e