Implement basic crash detection and shutdown handling for out of process PPAPI plugins.
Currently when a crash is detected we just delete as much stuff as is
convenient, clear the plugin's backing store, and continue running. It does not
hook up a sad plugin page yet.

This adds a "proxy" interface for the proxy to tell the PPAPI backend
implementation in the renderer about proxy-related stuff (like the plugin
crashing).

This also implements keeping the process alive for a period of time so we can
re-use the same process.
Review URL: http://codereview.chromium.org/6493004

git-svn-id: http://src.chromium.org/svn/trunk/src@74965 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: d69dd34d88f430d46352e8b6db6e267c833a39a8