views transformation: First cut.
This is a first cut at the transformable views. Skia transformation matrices are used to apply transformation (rotate, scale, translation). Support for clipping is also added (but not through the matrix). Currently, the transformation is applied only for painting. A future CL will apply the transformation for various events.

BUG=none
TEST=ViewTest.TransformPaint

Review URL: http://codereview.chromium.org/6500008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@75023 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d0f56b7ee14acf6cb7692e7a8ac91814fa00a04e