Autofill cleanup of labels
This removes remaining vestiges of profile and credit card labels as a means for identification.  Labels do remain in the profile and credit card classes but are private, not persisted to the database, and created upon demand.  This also cleans up the web_data_service and web_database methods that used labels as identifiers.

BUG=72929
TEST=WebDataServiceAutofillTest.*:WebDatabaseTest.* and miscellaneous others.

Review URL: http://codereview.chromium.org/6480073

git-svn-id: http://src.chromium.org/svn/trunk/src@75028 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 3f7d4c0f01f588542f4a3ba083aa6ff032ed2dc6