Fix is_convertible for windows.
It turns out that if you nest a template inside the initializer list of another template, the namelookup in windows gets all funky.  In this case, it couldn't tell that the found name was a function.

Luckily, we could just move the template parameters one level deeper.

BUG=none
TEST=unittests & trybots.

Review URL: http://codereview.chromium.org/6537001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@75194 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 1432236e20f20839da695ffcf3b8a144da64d48c