Call method CheckTitle to verify title of the active tab.
After opening a new tab, the tab may not be ready to read the title. The method CheckTitle checks title within the certain timeout.
It may not a good solution but workable and existing code already uses it, so I think we use it to fix this bug.


BUG=46781
TEST=UnloadTest.BrowserCloseBeforeUnloadOK.

Review URL: http://codereview.chromium.org/6544015

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@75386 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 3ec12af7d96406361dfe88f16b121447e937f34b