Call method CheckTitle to verify title of the active tab.
After opening a new tab, the tab may not be ready to read the title. The method CheckTitle checks title within the certain timeout.
It may not a good solution but workable and existing code already uses it, so I think we use it to fix this bug.


BUG=46781
TEST=UnloadTest.BrowserCloseBeforeUnloadOK.

Review URL: http://codereview.chromium.org/6544015

git-svn-id: http://src.chromium.org/svn/trunk/src@75386 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: c339ef2f33d408241ace22558438b04d39305887