Implement the filesystem proxy. This allows the FileRef tests (the ones which don't use FileIO which isn't don yet) to pass in the proxy.
Hook up the code from the URLLoader that downloads to a file. This allows all
URLLoader tests to pass in the proxy.

Change code in dispatcher that zeros out the array to take into account
padding. It was only zero-filling half of the array since sizeof(enum) * # elts
seems to be half as large as the actual array due to padding on 64-bit systems.

Make the aborted completion callbacks run asynchronously. This was caught by
one of the file ref tests. We really need a system for doing this better, but
I don't want to do that in this patch.

TEST=ppapi_tests run under proxy
Review URL: http://codereview.chromium.org/6543028

git-svn-id: http://src.chromium.org/svn/trunk/src@75566 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: d1f3ed7e5ceb0c249f636b70a79c2a13f8f05917