chromeos: Check ALSA return values in mixer code.
We're seeing a NaN value sometimes get saved (incorrectly)
to the prefs file in the volume setting, which results in
the prefs file being unparseable the next time Chrome
starts, which results in us going through OOBE again.  The
prefs code ought to write NaNs correctly so that they don't
corrupt the file, but in the meantime, this change checks
the return values from a bunch of ALSA functions (my best
theory as to how the NaN is creeping in) and for good
measure, also maps NaN volume values to the minimum volume.

BUG=chromium-os:12229
TEST=built and ran it (i haven't been able to repro the problem)

Review URL: http://codereview.chromium.org/6562001

git-svn-id: http://src.chromium.org/svn/trunk/src@75698 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: b5ffa11f693f8e385a6ce8e6d6b57d2cd7a2bfac