WebUI Prefs: Don't use display:block for labels
Separates our labels into a div and a label, so that labels are no longer display:block. Making labels display:block turns out to be undesirable because it makes the clickable area for the control the full width of the pane, rather than just the control and the label text.

BUG=69792
TEST=Click to the right of a checkbox or radio button's label; it should not active the control.

Review URL: http://codereview.chromium.org/6541086

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@75772 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 661c400a9d27902e20ffb859f91f9502cfd2b01d