Autofill crasher Chrome:+Crash+Report+-+Stack+Signature:+WebCore::Node::renderBoxModelObject()
Attempt at fix for new crasher specific to 11.0.680.0.  The Autofill code on the
renderer side has not changed in a couple weeks.  Upon inspection of
WebCore::Node::renderBoxModelObject() it would seem that the |this| pointer is
NULL.  This change adds an isNull() check on the node in question, but this
doesn't address the source of the issue.  Will follow-up with QE to bisect and
isolate if possible.

BUG=73895
TEST=observe crash reports on canary

Review URL: http://codereview.chromium.org/6577009

git-svn-id: http://src.chromium.org/svn/trunk/src@75826 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 0cea547db71dcab107039cf164a178c1332b0542