Added TranslateVariablesInPolicy to the policy toolkit.
This function was implemented for mac and win in the chrome_main_xxx files because it was used only by the UserDataDir policy until now. However we need this in the DownloadDirectory policy too so I extracted in in a more
global way in the policy::path_parser namespace.

BUG=None
TEST=unit_tests:PolicyPathParser.*

Review URL: http://codereview.chromium.org/6469081

git-svn-id: http://src.chromium.org/svn/trunk/src@76045 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 84238e9225bfc285bab9a1596bd58f52bc5e69e8