[Reason for revert: might be a factor in XP-only failures UITestCanLaunchWithOSMesa.  Will unrevert if proven innocent.]
Revert 76075 - With this CL, GPU blacklist auto update from the web (before it ships together with chrome) is implemented.  This allows us to blacklist bad GPU/drivers as soon as we discover them.

Note that this patch does not turn the auto update on.  We will turn it on in a separate CL.

BUG=68802
TEST=build bots all green

Review URL: http://codereview.chromium.org/6469094

TBR=zmo@google.com
Review URL: http://codereview.chromium.org/6599020

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@76122 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 54b9b2a6fd6006d7ef17b77a957382eb1cae603e