Add a noise indicator to the speech bubble volume indicator.
The noise indicator is drawn as a light blue area at the beginning and if there
was clipping that is denoted with a red area at the end of the meter. The noise
level comes from the endpointer -> SpeechRecognizer -> SpeechInputBubbleController -> SpeechInputBubble
hence a bunch of volume setting methods are updated with the new parameter.

I have also added a new utility method to SpeechInputManager to invoke the platform
provided microphone settings UI, this will be used in the next CL which contains
windows, mac and linux specific UI changes.

BUG=69886
TEST=manual, invoke speech input and check the bubble volume indicator to see background noise and clipping.

Review URL: http://codereview.chromium.org/6597071

git-svn-id: http://src.chromium.org/svn/trunk/src@76395 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 4f5a5eda59e1071b739fe34b49ef18425807d432