Adds PrerenderBrowserTest.PrerenderNoSSLReferrer, which checks that there is no referrer when an SSL page prerenders a non-SSL page.
Note that this patch is most of patch http://codereview.chromium.org/6580041
(reverted http://src.chromium.org/viewvc/chrome?view=rev&revision=76304).
The only difference is the redirect test is not modified to use plugins.

After the patch, the buildbots were periodically failing
PrerenderBrowserTest.PrerenderRedirect with lock errors (2 times out of 10+
runs).  See, e.g.
http://build.chromium.org/p/chromium/builders/Linux%20Tests%20x64/builds/5976/steps/browser_tests/logs/PrerenderRedirect
(Note that the percents in the URL are incorrectly escaped by the
codereview page, so the URL has to be copied and pasted to be used).

I believe the issue is that the other prerender browser test that uses plugins
is not shutting down the plugin process properly.  I want to see if I can
reproduce the issue locally and fix it, not because the change I reverted
is so important, but because if I'm right, it could be a cause of future
test flake.

BUG=73170
TEST=PrerenderBrowserTest.PrerenderNoSSLReferrer

Review URL: http://codereview.chromium.org/6591068

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@76542 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: e235eb907c40f223623cc3fecb2295d397c2bdd8