Redo the speech bubble layout on windows.
The volume indicator is moved up, 'Speak now' text comes in the middle and when recognition
is in progress we show only the spinner with no text. When showing error messages, include
a hyperlink to launch the microphone settings UI. Also includes a bunch of spacing changes
to look like in the mocks.

Also changed usage of BrowserThread::PROCESS_LAUNCHER to BrowserThread::FILE per John's suggestion in an earlier review.

BUG=69886
TEST=manual, start speech input on windows and verify the bubble UI looks similar to the mocks.

Review URL: http://codereview.chromium.org/6604006

git-svn-id: http://src.chromium.org/svn/trunk/src@76556 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: bc953a29cdc87d7450bc2ccf7cd9163758a293e6