Record histograms for Prerender control group.
The histograms were not being recorded since PrerenderPLTRecorder was checking for existence of a PrerenderManager in the current Profile, and one is not present for the control case.

I fixed this by making the PLT recording functions static: they already had no instance-specific dependencies, so this works pretty easily.

BUG=None
TEST=Force Chrome to be in prerender control group [requires modification to prerender_field_trial]. Do a query for "Google Trends" and click on it. about:histograms should have Prernder control page load times and windowed page load times.

Review URL: http://codereview.chromium.org/6597109

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@76602 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 397c9e78f517eb2902f208ca6ac0077a1c8f1bff