Record histograms for Prerender control group.
The histograms were not being recorded since PrerenderPLTRecorder was checking for existence of a PrerenderManager in the current Profile, and one is not present for the control case.

I fixed this by making the PLT recording functions static: they already had no instance-specific dependencies, so this works pretty easily.

BUG=None
TEST=Force Chrome to be in prerender control group [requires modification to prerender_field_trial]. Do a query for "Google Trends" and click on it. about:histograms should have Prernder control page load times and windowed page load times.

Review URL: http://codereview.chromium.org/6597109

git-svn-id: http://src.chromium.org/svn/trunk/src@76602 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: a3e2b30ce4706e56c0cab92fb8c7bd7be0cd208d