Redo the speech bubble layout on GTK.
The volume indicator is moved up, 'Speak now' text comes in the middle and when recognition
is in progress we show only the spinner with no text. When showing error messages, include
a hyperlink to launch the microphone settings UI. Also includes a bunch of spacing changes
to look like in the mocks.

BUG=69886
TEST=manual, start speech input on GTK and verify the bubble UI looks similar to the mocks.

Review URL: http://codereview.chromium.org/6617001

git-svn-id: http://src.chromium.org/svn/trunk/src@76736 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 6a7406389aa5f80065b64d1d12abc603aad1d135