Redo the speech bubble layout on mac.
The volume indicator is moved up, 'Speak now' text comes in the middle and when recognition
is in progress we show only the spinner with no text. When showing error messages, include
a hyperlink to launch the microphone settings UI. Also includes a bunch of spacing changes
to look like in the mocks.

XIB changes: Added a NSButton widget, changed the cell class to HyperlinkButtonCell and linked it to the obj-c code for handling the click event.

BUG=69886
TEST=manual, start speech input on Mac and verify the bubble UI looks similar to the mocks.

Review URL: http://codereview.chromium.org/6594122

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@76751 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: a984ac15b6312663a28439f7d9a84c3e9f6bdc7e