ChromeFrame would fail to upload POST data to the server if the webserver requested NTLM authentication. This is due to a bug in urlmon on IE6 and IE7 which manifests itself when the post data is passed to urlmon as an IStream.
Fix is to pass in the uploaded data as a HGLOBAL. We always pass in a copy of the HGLOBAL
which points to the posted data to urlmon. This is to have it accessible for reissuing
navigation requests which target downloads.

Fixes bug http://code.google.com/p/chromium/issues/detail?id=62687

BUG=62687
TEST=manually at this point. As we need a server which supports NTLM authentication like IIS.

Review URL: http://codereview.chromium.org/6603006

git-svn-id: http://src.chromium.org/svn/trunk/src@76880 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 0d31f6b8962ca89b11ee7c54880d84e80e48c43f