Add an optional 'page_index' preference to apps and 'setPageIndex' message which can be used to set it.
This doesn't actually modify the behavior of the NTP whatsoever.  It's a small extension which will be used by a future alternate NTP front-end that shares the same WebUI back-end code.  Any NTP front-end that doesn't support pagination of apps will simply ignore this.

This also includes a simple unit test to verify the functioning of this preference.

BUG=73437
TEST=new test in unit_tests ('ExtensionPrefsPageIndex')

Review URL: http://codereview.chromium.org/6546017

git-svn-id: http://src.chromium.org/svn/trunk/src@76916 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: c7bcaecfd5e907d71d9ccf698e5361aa37ab01dc