fix display of filenames in file:/// URLs
(retry of previous code with tests fixed)

The code attempted to shorten file:/// URLs to just the filename
when displayed as the title of a page.  But that appears to have
regressed sometime in the past.  This shortening is consistent with
how we display the title of images (which are like "foo.png (123x456)".)
Chrome does a poor job of displaying longer titles (most of the
tab title ends up being "file:///C:/" anyway).

In any case, using a FilePath to get the filename from a URL may not
have even worked on Windows, where the path separator is a backslash.

It appears Glen wrote the original code, Brett may be the one to have
regressed it in a refactor, and I probably broke it worse in a FilePath
refactor.

BUG=69467
TEST=load a text file via a file: URL; tab title is just the file name

Review URL: http://codereview.chromium.org/6591127

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@77008 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 9c34ca02ae0bf72cc3603d7374cf24a4ae439afa