Fix extra "&" in back/forward menu on Mac
All "&" characters in the back/forward menu would show up as "&&" on the Mac.

The problem was that Windows uses "&" as an accelerator in menus. We were escaping all "&" by replacing them with "&&".

r36530 added code to the GTK menu to strip the accelerator. Unfortunately those util functions are in GTK only files. Instead of copying that approach I'm simply #ifdefing out the code that adds the extra "&" characters.

BUG=17660
TEST=Verified that back/forward menus look correct.


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@77087 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 10fd4f37536b04adbd650bac42a1836757f8175a