Add link to activate verizon in offline page. This CL also update the page itself to match new mock from kenmoore.
Fix a crash when reloading offline page. It calls DontProceed when reloaded, so I shouldn't have Release reference on OnRequestClosed.

There are a couple of difference between the page and mock and will continue working on it for R12, but this should be good enough for R11. (or we can do minor update in the branch)

BUG=chromium-os:12286
TEST=manual: on chromeos, disconnect network and load www.google.com. Offline page now has a link to activate verizon 3g.

Review URL: http://codereview.chromium.org/6625020

git-svn-id: http://src.chromium.org/svn/trunk/src@77172 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 3e2762cfa83d6de55110d5b70a170b399ede8281