Add link to activate verizon in offline page. This CL also update the page itself to match new mock from kenmoore.
Fix a crash when reloading offline page. It calls DontProceed when reloaded, so I shouldn't have Release reference on OnRequestClosed.

There are a couple of difference between the page and mock and will continue working on it for R12, but this should be good enough for R11. (or we can do minor update in the branch)

BUG=chromium-os:12286
TEST=manual: on chromeos, disconnect network and load www.google.com. Offline page now has a link to activate verizon 3g.

Review URL: http://codereview.chromium.org/6625020

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@77172 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 5682b9f949d0806964df3ba2488822831d20f186