Make the toast more reliable and less dangerous
The current toast (re-activation) experiment suffers from the following problems:

1- False negative: the toast can be triggered even if chrome is being used
2- Nested message loop crash: If chrome tries to run while the toast is being shown, it crashes

This change addresses both conditions. Before running the toast we leverage the
process_singleton to detect another instance. If so we abort with "no thanks".

When showing the toast and its nested message loop we take the process_singleton lock which
inhibits dispatch of the wm_copydata message which spawns new tabs in the current process if
we try to span a new tab while in the nested message loop we crash.

BUG=32026
TES=see bug


Review URL: http://codereview.chromium.org/6625037

git-svn-id: http://src.chromium.org/svn/trunk/src@77214 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: e6b2ca8f40f20021957c6851c4ec31ba7579a853