Adding an image to avoid future try server failure.
This is for the upcoming redesign changelist for the
redesign of the Manage Extensions page (as a Settings
page).

Try servers tend to fail (at least have in the past)
if a binary file was included in the changelist
being 'tried'.

This should avoid that problem.

BUG=52447
TEST=None
TBR=mwichary

Review URL: http://codereview.chromium.org/6652021

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@77471 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: c68d03d3bca3d3556fc795161f548265109f8285