12 (twelve) (pronounced /ˈtwɛlv/ ( listen)) is the natural number following 11 and preceding 13.
The word "twelve" is a native English word that presumably arises from the Germanic compound twa-lif "two-leave", meaning that two is left after one takes away the base, ten. This compound meaning may have been transparent to speakers of Old English, but the modern form "twelve" is quite opaque. Only the remaining tw- hints that twelve and two are related. Etymology (Weekley, Skeat) suggests that "twelve" (similar to "eleven") consists of two parts, the first meaning "two" and the second "leftover", so a literal translation would yield "two remaining [after having ten taken]".

A group of twelve things is called a Duodecad. The ordinal adjective is duodenary, twelfth. The adjective referring to a group consisting of twelve things is duodecuple.
The number twelve is often used as a sales unit in trade, and is often referred to as a dozen. Twelve dozen are known as a gross. (Note that there are thirteen items in a baker's dozen.)

As shown below, the number twelve is frequently cited in the Abrahamic religions and is also central to Western calendar and units of time.

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@77670 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 877fdaa6cd82431969d5ff7ed800ae1a738373c6