Add in DOMBrowser
This is the second CL for a sequence of CLs that will add in the display infrastructure that I have built for the DOM Login and touch keyboard. I have created a reference CL for what the end product of this sequence is intended to be, so that I can give context for the CL under review. This can be found at: http://codereview.chromium.org/6577003/

This CL adds in DOMBrowser to touchui==1 and chromeos==1 builds. This class is a subclass of Browser. It exists mostly to allow the DOMBrowserView and DOMBrowserFrame classes to be used, which will come in following CLs. This CL also extends the guard arround set_window in Browser to allow it to be used in touchui && chromeos builds. This is done to allow DOMBrowser access to the method, since window_ is private not protected. There are also additional checks in Browser added for location_bar being NULL, which are used when DOMBrowserView is the BrowserView, since the location bar does not exist for it.

Patch from Ryan Harrison <rharrison@chromium.org>

BUG=none
TEST=manual

Review URL: http://codereview.chromium.org/6626014
Patch from Ryan Harrison <rharrison@chromium.org>.

git-svn-id: http://src.chromium.org/svn/trunk/src@77830 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 296393128b1af3e0f9ef5c8fdeb3852ffc042232