Don't drain response bodies for SpdyHttpStreams.
This should fix the specific instance of crbug.com/75657, although I can't reproduce the bug, so it's possible it may instead push the bug later down the pipeline rather than completely fixing it. I believe it should completely fix it though.

This is a partial reland of r77399, which was reverted in order to faciliate this partial reland so I can merge a smaller change into branch 696.

BUG=75657
TEST=none

Review URL: http://codereview.chromium.org/6680009

git-svn-id: http://src.chromium.org/svn/trunk/src@77893 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: d10c1f9db193885cafaa60e48c54df3e87fa6602