[Mac] Tighter restrictions on NSException raises.
On the crash server, almost all NSInvalidArgumentException cases are
due to sending a bad selector to an object.  Most likely cause is a
stale pointer which has been reallocated to another object, or memory
stompers.

Also start tracking uncaught NSInternalInconsistencyException, which
is the most popular exception on the crash servers.

BUG=24462
TEST=none

Review URL: http://codereview.chromium.org/6698002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@78249 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 03d009347e99fd10d2abf0daec784b1ca0a22c06