Introduce gpu_trace_event for gpu performance analysis.
This changelist is a lightweight version of issue 6551019, but with
the code change confined to the gpu subsystem.

The intent of this change is to enable forward progress on gpu tracing.

Work on the aformentioned issue, i.e. fusing gpu_trace_event and trace_event, will continue.

BUG=
TEST=

Review URL: http://codereview.chromium.org/6691013

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@78300 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 46c2959993646fd621eb3d76cc780b9a19a9aa36