Make URLRequestHttpJob the only URLRequestJob concerned with FilterContext.
In order to allow this, provide a new factory method in Filter to return a
GZipFilter (used by URLRequestFileJob and URLRequestJobTrackerTest), and
invert control so that URLRequestJobs are responsible for constructing filters.

This is one step away from removing FilterContext as a base class of URLRequestJob,
which will be tackled in a followup change.

BUG=none
TEST=net_unittests

Review URL: http://codereview.chromium.org/6674042

git-svn-id: http://src.chromium.org/svn/trunk/src@78489 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 99d2a641d4c416b9ffb72efb064bbeb99c2442f4