Make URLRequestHttpJob the only URLRequestJob concerned with FilterContext.
In order to allow this, provide a new factory method in Filter to return a
GZipFilter (used by URLRequestFileJob and URLRequestJobTrackerTest), and
invert control so that URLRequestJobs are responsible for constructing filters.

This is one step away from removing FilterContext as a base class of URLRequestJob,
which will be tackled in a followup change.

BUG=none
TEST=net_unittests

Review URL: http://codereview.chromium.org/6674042

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@78489 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: b82310018a3311f91e18f24e12a89a74ccd2e819