Stop subclassing FilterContext in URLRequestJob.
Create a wrapper for URLRequestHttpJob that subclasses FilterContext,
and pass an instance of this facade to Filter::Factory().

Reduce the scope of URLRequestJob's interface as much as possible, moving
methods into URLRequestHttpJob or its facade as appropriate, and making
methods that remain in URLRequestJob non-virtual where possible.

BUG=none
TEST=net_unittests,try bots

Review URL: http://codereview.chromium.org/6677104

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@78576 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: a42ed14e9d3dfa887b6dd7f88ea351dc096fb419