Autofill database migration to clean up bogus profiles.
Adds a |WebDatabase| migration to clean up old profiles.  The profiles are merged in much the same way as when aggregated using form submission.  Profiles that are identical to another, subsets of another, or contain invalid or incomplete email, state, zip, or address get filtered out.

Additionally, the filtered profiles are remembered in a "trash can" mechanism to later reflect these changes in Sync, so that they don't reappear unintentionally.

BUG=65625
TEST=PersonalDataManagerTest.*:AutoFillMergeTest.DataDrivenMergeProfiles:WebDatabaseMigrationTest.*:WebDatabaseTest.*

Review URL: http://codereview.chromium.org/6676031

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@78589 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: db5537f06c924c48fef5f10e8481113415bfc8eb