Autofill database migration to clean up bogus profiles.
Adds a |WebDatabase| migration to clean up old profiles.  The profiles are merged in much the same way as when aggregated using form submission.  Profiles that are identical to another, subsets of another, or contain invalid or incomplete email, state, zip, or address get filtered out.

Additionally, the filtered profiles are remembered in a "trash can" mechanism to later reflect these changes in Sync, so that they don't reappear unintentionally.

BUG=65625
TEST=PersonalDataManagerTest.*:AutoFillMergeTest.DataDrivenMergeProfiles:WebDatabaseMigrationTest.*:WebDatabaseTest.*

Review URL: http://codereview.chromium.org/6676031

git-svn-id: http://src.chromium.org/svn/trunk/src@78589 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 48b3172b27bd7949cb866a09368e39e0fee710ba