WebUI Prefs: Switch adding a homepage to an inline field with autocomplete
This makes adding a homepage match the mocks and the other WebUI pref lists, adding autocomplete so there's no loss of usability from removing the "recent pages" list.

This refactors inline-editable lists to understand the concept of the placeholder row since it's used in multiple places. It also tweaks the autocomplete list to be more reusable.

BUG=69689
TEST=Add a page to the start list. It should be inline, and should suggest urls from history.

Review URL: http://codereview.chromium.org/6670061

git-svn-id: http://src.chromium.org/svn/trunk/src@78772 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 500c966de833c88647dc62d7137cabf8a09c28b2