Inform user to reload page after chagning cookie settings from in-page dialog.
NOTE: the png file is not final yet, waiting for chrome ui folks to provide
final vesion.  However, this does not change the code, so please review.
Thanks.

BUG=63645
TEST=Block all cookies and load a page that uses cookies.  Click on the broken
cookie icon in the omnibox and open the in-page cookie dialog.  Change a
blocked to allow, or block an allowed cookie.  After clicking the Close button
an infobar appears with a warning.  Make sure the reload button works.
Review URL: http://codereview.chromium.org/6676041

git-svn-id: http://src.chromium.org/svn/trunk/src@78829 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 46488915c16f1f0857bb4bd4de861cae24ea393a