Add in DOMBrowserView and Frame related classes
This is the final CL for a sequence of CLs that will add in the display infrastructure that I have built for the DOM Login and touch keyboard. I have created a reference CL for what the end product of this sequence is intended to be, so that I can give context for the CL under review. This can be found at: http://codereview.chromium.org/6577003/

This CL adds in DOMBrowserView, DOMBrowserFrame, and related classes to touchui==1 and chromeos==1 builds. DOMBrowserView classes inherit from the equivalent chromeos::BrowserView classes. DOMBrowserFrame classes inherit from TouchBrowserFrame classes, since it has the keyboards implemented in it. In addition to the direct parent I have had to make changes to other classes in the hiearchy as needed to ease implementation. Additional checks and refatoring have been done to get the build to work and minize the size of this CL. I have actively chosen to not refactor chromeos::BrowserViewLoayout out of the file it current resides in since DOMBrowserViewLayout wouldn't inherit much from it if I subclassed it.

Patch from Ryan Harrison <rharrison@chromium.org>

BUG=none
TEST=manual


Review URL: http://codereview.chromium.org/6692001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@79018 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 029f54499be575dcd549d962c6e4f828bb9b0367