Fix for "Mouseovers follow the cursor even when there's a find bar in the way"
Perform a hit test in the render widget host view to check if the cursor is
actually on top of a different view - such as the find bar. If it is and that
view has a method -[(BOOL)nonWebContentView] that returns YES, then ignore
the event in the render widget host.

Change FindBarView to have the aforementioned method that returns YES.

Patch by Alexei Svitkine <asvitkine@chromium.org>

BUG=22191
TEST=Leave the content area by mousing over a visible find bar. Mouseovers
     from the content area should extinguish, and no masked by the find bar
     should be triggered.
R=mark@chromium.org
Review URL: http://codereview.chromium.org/6676094

git-svn-id: http://src.chromium.org/svn/trunk/src@79438 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 3ed56e7d87bf870ca3f2cf1b370f4f17fe658dfb