Disable HTTP auth schemes on permanent errors.
The underlying implementation for the Negotiate authentication
scheme might return error codes that indicate error conditions
that we are unlikely to recover from.  If we see those, then
treat these as permanent errors and disable the auth scheme for
the rest of the transaction.  We were already doing this partly
for cases where there are no default credentials.  This patch
extends the behavior to additional error codes.

BUG=49950
TEST=net_unittests --gtest_filter=HttpAuthControllerTest.PermanentErrors

Review URL: http://codereview.chromium.org/6745018

git-svn-id: http://src.chromium.org/svn/trunk/src@79848 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: aff722f4728947539e03dafb47ea52e7308017b4