Refactor and rewrite of the password and passphrase sync integration tests
The sync password and passphrase integration tests have been in need of an overhaul due to the fact that the implementation of password and passphrase sync has changed significantly over the past couple months.

This patch updates the LivePasswordsSyncTest class, optimizes the wait methods in ProfileSyncServiceHarness related to passphrases, and simplifies the state machine used by SetupSync(). It also contains a partial rewrite of the password integration tests based on the updated framework.

BUG=59867, 67862, 77956
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/6698072

git-svn-id: http://src.chromium.org/svn/trunk/src@79892 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 2c62b11330f48db848ca92cccba6865e84502090