Notify child widgets before and after a reparenting.
Attaching a native view to a NativeViewHostWin may cause that
view to be disassociated from a view hierarchy with a different
FocusManager.  Notifying the widgets that they are about to be
removed from their current parent gives them the opportunity to
disassociate themselves from their current FocusManager.

BUG=77447
TEST=none


Review URL: http://codereview.chromium.org/6776025

git-svn-id: http://src.chromium.org/svn/trunk/src@80096 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 5eb2af3cf162b0d660de66780ee0cf052a5455e1