Implement CrashUploadListWin to enable chrome://crashes on Windows.
Added new class CrashUploadListWin, subclass of CrashUploadList which
has the Windows platform-specific logic to read crash logs from Windows
Event Log.

Added static CrashUploadList::Create() factory method which returns an
instance of CrashUploadListWin on Windows and an instance of
CrashUploadList on other platforms.

Also implemented the max_count parameter in
CrashUploadList::GetUploadedCrashes(), which was part of
the interface but was ignored by the code.

BUG=73468
TEST=Open chrome://crashes under Windows after one or more
     crash reports were uploaded. See the crash reports listed.


Review URL: http://codereview.chromium.org/6771021

git-svn-id: http://src.chromium.org/svn/trunk/src@80165 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 4a47b0ed80a76edbe35f2aa726ae91456a877d6a