Fix resource destruction in proxy
This ensures that the resource on the plugin side is destroyed before we send
the message to the host, so that it has a chance to do proper cleanup.

Also, fix Surface3D destruction that could cause a write-after-free.

BUG=none
TEST=go to youtube with out-of-process pepper flash. click on fullscreen.
observe no hang, no crash

Review URL: http://codereview.chromium.org/6771042

git-svn-id: http://src.chromium.org/svn/trunk/src@80188 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: a91f7b39384cb877bad4000c9c24576002de1385