Make <all_urls> and file:///* in permissions trigger "Allow file access"
Move ExtensionWantsFileAccess static functions to Extension::wants_file_access
which is set as part of the Extension initialization process.

When file:/// access is granted/revoked, instead of storing that in a bit on the
UserScript (which didn't help with extension host permissions), we use that
to filter down the list of valid schemes for permissions/content script match
matterns. That way we don't have to manually exclude file:///* patterns at
various call sites (e.g. in UserScriptSlave).

BUG=76705
TEST=unit_tests
R=aa@chromium.org

Review URL: http://codereview.chromium.org/6772022

git-svn-id: http://src.chromium.org/svn/trunk/src@80213 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 6089682c4b8371061d69df2366914289d4c80b81