Fix DCHECK() in infobar animation.
The core of the problem here is that painting does not always happen synchronously with layout, and the old code was not always forcing re-layout to occur if the animation progressed in the meantime.  This change forces re-layout to occur any time the animation progresses or the target height is changed, and makes that the one time when the tab and bar heights are recalculated -- all other places then use the calculated values.

This also tweaks which functions are in the cross-platform versus platform-specific code as well as making a few other small cleanups.

Finally, this re-enables a test on Mac that seems to be passing now.

BUG=60990, 75451
TEST=none
Review URL: http://codereview.chromium.org/6788014

git-svn-id: http://src.chromium.org/svn/trunk/src@80272 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 2c607cd13a2b4b0b9a484d995e5975fdeb5bea96