Initial file manager dialog implementation.
Here's the initial file manager implementation and development harness.

After applying this patch you should be able to visit file://path/to/chrome/src/chrome/browser/resources/file_manager/harness.html to see dialog in the harness.

You'll have to start chrome with
`chrome --allow-file-access-from-files --unlimited-quota-for-files`
Then click on the 'Choose File' button at the bottom of the harness to add files to the filesystem used by the dialog.  Watch the js console for status messagesa bout the import, and reload the page when the import is complete.  The 'Choose File' dialog wonn't allow you to import a directory structure, so you may want to use bin/squashdir.py to translate a directory tree into something that can be reconstituted by the harness.

This code only implements the 'select one file' mode at the moment, the other modes are coming soon.

Thumbnails are currently displayed using a data: url for the entire image file.  This will change when soon when we have a thumbnail data source, or a js thumbnailer.

Icons and preview images are only placeholders until the real assets are available.

BUG=
TEST=

Review URL: http://codereview.chromium.org/6723010

git-svn-id: http://src.chromium.org/svn/trunk/src@80335 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 7a45245f3c0a41162167f2dba8da2abbcf78938c