Add parser and test files
This CL adds the PLY based parser, parser test files, and makes some additional
cleanup to the lexer.  The parser is part of a generator which is still under
development.  The parser can be run standalone which works as a test mode
verifying success and failure tags it find in the sources embedded as 
comments.  Adding the switch  --output will print the AST.


R=sehr@google.com
BUG=77551
TEST=python idl_parser.py test_parser/*.idl
Review URL: http://codereview.chromium.org/6754001

git-svn-id: http://src.chromium.org/svn/trunk/src@80933 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 6a7dc6649bef63c1f23fc3becda152a6c490f8f2