Minor html fix and changed the size of the control links in interim login screen
I have been having issues with clicking the targets on this page during development do to other issues. Since this is not intended to be the final screen I just increased the size of the text so I don't accidentally click the wrong link. This caused it to be obvious that one of my divs wasn't closed correctly, so I fixed that also.

Patch from Ryan Harrison <rharrison@chromium.org>

BUG=none
TEST=manual


Review URL: http://codereview.chromium.org/6801048

git-svn-id: http://src.chromium.org/svn/trunk/src@81155 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 95df6a61bc73859f9de2c17a459ec96cee76db43