Don't flush trace events until tracing ends.
Flushing trace events really disrupts the traced process, so don't flush events until tracing has finished. The events would be stored in memory after they were flushed anyway, so this doesn't change much. For this to work we must make sure to flush the events in small batches, so the send doesn't reject them.

Also add a limit of around 500000 on the number of trace events, so we don't run out of space due to them.

BUG=
TEST=

Review URL: http://codereview.chromium.org/6826043

git-svn-id: http://src.chromium.org/svn/trunk/src@81210 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 1da9d188c54112c289bad51af40b334a321db7d5