Mac: Let chrome coexist with the Chinese Handwriting IME.
The renderer sandbox breaks the Chinese Handwriting shortcut while Chrome
is running, even when Chrome is not active: In a nutshell, Carbon tries to load
the IME in every process, and if that fails, it uninstalls the toggle shorcut. It
fails in the renderer process due to the sandbox. See the bug for details.

Prewarming (patch set 1) doesn't work because that opens up clipboard
access to the renderer – the IME doesn't load correctly if it doesn't have
clipboard access.

dyld __interpose based interposing (patch set 2) doesn't work because
TISCreateInputSourceList and its broken caller are in the same
framework (HIToolbox).

Hence, use binary rewriting to return an arbitrary array with size > 0 to make
HIToolbox believe that the IME loads correctly, so that it doesn't uninstall the
global handwriting keyboard shortcut (only on >= 10.6).

BUG=31225
TEST=Enable Chinese in the input menu. Make sure Chrome's not running.
Note that ctrl-shift-space opens and dismisses the Handwriting IME
(you need a multitouch trackpad for that). Open Chrome, note that the
IME still works.
(If you run Chrome without this patch, the IME will be broken until you do
  ps aux | grep Chinese
and then
  kill $CHINESE_HANDWRITING_PID
and then toggle the IME once through the Menu Extra menu.)
It is a good idea to have a Terminal open during testing, because killing
the task is the only way to dismiss the IME with the keyboard without this
CL.

Committed: http://src.chromium.org/viewvc/chrome?view=rev&revision=81308

Committed: http://src.chromium.org/viewvc/chrome?view=rev&revision=81332

Review URL: http://codereview.chromium.org/6801056

git-svn-id: http://src.chromium.org/svn/trunk/src@81500 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 4d2cbec78a3e180aa981e5f4552fb763a8806b03