Whitelist access to the disk in URLRequestResourceBundleJob::GetData.
This function is run on the IO thread, which means it shouldn't touch the disk,
but it does.  There's already a bug on file for fixing this issue (which covers
this instance and the same instance in some of the other URLRequestJob
subclasses).  I'm adding ScopedAllowIO in this patch to silence the DCHECK
while we work on the bug.

BUG=59849
Review URL: http://codereview.chromium.org/6851026

git-svn-id: http://src.chromium.org/svn/trunk/src@81656 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: af0b41e8ab2cf854c9f705452a45de97c5f6f74c