Split WGL implementations of *GLContext in GLContextWGL and *GLSurfaceWGL.
Surfaces are independent of contexts in GL. To facilitate sharing of surfaces between processes, I have separated them from the notion of contexts because contexts cannot be shared between processes.

I started with EGL in r81512. This is the same thing for WGL.

GLContextEGL still has a pointer to a surface and still has some surface specific operations that just forward through to it. Once I have refactored all the GLContext implementations in this way, I will remove these pointers and the surface specific opertations.

There will not be "view" and "offscreen" GL contexts. Rather there will be a single context type for each backend which can be made current with a surface that directs output either to a view or offscreen surface.
Review URL: http://codereview.chromium.org/6864002

git-svn-id: http://src.chromium.org/svn/trunk/src@81807 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: d2ba658ab1b969ceac95d3b028269e27bcbf3810