New automation hook to set the app launch type (e.g., window vs. tab).
This hook allows the launch type for a specified app to be set as either
a regular tab, a pinned tab, a fullscreen tab, or a tab in a new window.
Sample PyAuto tests are included that exercise this hook.  The GetBrowserInfo
automation hook is also slightly modified to return information about
whether each existing tab is pinned or not.

BUG=75205
TEST=None.

Review URL: http://codereview.chromium.org/6852005

git-svn-id: http://src.chromium.org/svn/trunk/src@81808 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 4e5b40b8cb603235a2ce40a56e56abd2d37f5c5d