Support PPB_Flash_ModileLocal blocking calls from background threads. This is a temporary stopgap to allow these specific functions to be used until the entire proxy can support threading.
There are two functions that set up and clean up the routing information in
the interface. These are necessary because the global dispatcher routing
information is not threadsafe.

When used from the main thread, the code works the same as before. When used
from a background thread, we create a sync message, send it to the I/O thread,
and block the sending thread until the reply is received. There is no handling
of other incoming blocking messages (which are impossible), and no other
Pepper functions are supported from the background thread.

TEST=manual
BUG=none
Review URL: http://codereview.chromium.org/6875009

git-svn-id: http://src.chromium.org/svn/trunk/src@81814 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 84025266f28719bc33e9b289a6b40c7c9401e300